migrate or support more routes from src/app.js
