+++ server/src/app.js
@@ -26,74 +26,63 @@ when launching:
   both name and label are the same "name" from serverconfig
 */
 
-import serverconfig from './serverconfig'
-import express from 'express'
+import serverconfig from './serverconfig.js'
 import util from 'util'
 import url from 'url'
-import http from 'http'
-import https from 'https'
 import fs from 'fs'
 import path from 'path'
 import got from 'got'
 import lazy from 'lazy'
-import compression from 'compression'
 import child_process from 'child_process'
 import { spawn } from 'child_process'
 import { createCanvas } from 'canvas'
-import { stratinput } from '#shared/tree'
-import bodyParser from 'body-parser'
 import imagesize from 'image-size'
 import readline from 'readline'
-import jsonwebtoken from 'jsonwebtoken'
-import * as utils from './utils'
-import * as common from '#shared/common'
-import * as vcf from '#shared/vcf'
-import handle_study from './handle_study'
 import d3color from 'd3-color'
 import { stratify as d3stratify } from 'd3-hierarchy'
 import * as d3scale from 'd3-scale'
 import * as d3dsv from 'd3-dsv'
-import basicAuth from 'express-basic-auth'
-import * as termdb from './termdb'
-import { handle_tkbigwig } from './bw'
-import { handle_tkld } from './ld'
-import * as termdbbarsql from './termdb.barchart'
-import bedgraphdot_request_closure from './bedgraphdot'
-import bam_request_closure from './bam'
-import { mdsjunction_request_closure } from './mds.junction'
-import { gdc_bam_request } from './bam.gdc'
-import * as mds3Gdc from './mds3.gdc'
-import aicheck_request_closure from './aicheck'
-import bampile_request from './bampile'
-import junction_request from './junction'
-import bedj_request_closure from './bedj'
-import { request_closure as blat_request_closure } from './blat'
-import { mds3_request_closure } from './mds3.load'
-import { handle_mdssvcnv_expression } from './handle_mdssvcnv_expression'
-import { server_updateAttr } from './dsUpdateAttr'
-import * as mds2_init from './mds2.init'
-import * as mds3_init from './mds3.init'
-import * as mds2_load from './mds2.load'
-import * as massSession from './massSession'
-import * as singlecell from './singlecell'
-import * as fimo from './fimo'
-import { draw_partition } from './partitionmatrix'
-import mdsgeneboxplot_closure from './mds.geneboxplot'
-import { handle_mdssurvivalplot } from './km'
-import * as validator from './validator'
-import cookieParser from 'cookie-parser'
-import { authApi } from './auth.js'
-import { server_init_db_queries, listDbTables } from './termdb.server.init'
-import { versionInfo } from './health'
-import { decode as urlJsonDecode } from '../shared/urljson'
+import * as utils from './utils.js'
+import { stratinput } from '#shared/tree.js'
+import * as common from '#shared/common.js'
+import * as vcf from '#shared/vcf.js'
+import handle_study from './handle_study.js'
+import * as termdb from './termdb.js'
+import { handle_tkbigwig } from './bw.js'
+import { handle_tkld } from './ld.js'
+import * as termdbbarsql from './termdb.barchart.js'
+import bedgraphdot_request_closure from './bedgraphdot.js'
+import bam_request_closure from './bam.js'
+import { mdsjunction_request_closure } from './mds.junction.js'
+import { gdc_bam_request } from './bam.gdc.js'
+import * as mds3Gdc from './mds3.gdc.js'
+import aicheck_request_closure from './aicheck.js'
+import bampile_request from './bampile.js'
+import junction_request from './junction.js'
+import bedj_request_closure from './bedj.js'
+import { request_closure as blat_request_closure } from './blat.js'
+import { mds3_request_closure } from './mds3.load.js'
+import { handle_mdssvcnv_expression } from './handle_mdssvcnv_expression.js'
+import { server_updateAttr } from './dsUpdateAttr.js'
+import * as mds2_init from './mds2.init.js'
+import * as mds3_init from './mds3.init.js'
+import * as mds2_load from './mds2.load.js'
+import * as massSession from './massSession.js'
+import * as singlecell from './singlecell.js'
+import * as fimo from './fimo.js'
+import { draw_partition } from './partitionmatrix.js'
+import mdsgeneboxplot_closure from './mds.geneboxplot.js'
+import { handle_mdssurvivalplot } from './km.js'
+import { decode as urlJsonDecode } from '../shared/urljson.ts'
 
-export * as phewas from './termdb.phewas'
+export * as phewas from './termdb.phewas.js'
 
 export const tabixnoterror = s => {
 	return s.startsWith('[E::idx_test_and_fetch]') // got this with htslib 1.15.1
 }
 
 // cache
+// ??? ch_genemcount is not used anywhere ???
 const ch_genemcount = {} // genome name - gene name - ds name - mutation class - count
 const ch_dbtable = new Map() // k: db path, v: db stuff
 
@@ -101,7 +90,11 @@ export const features = serverconfig.features
 
 //////////////////////////////
 // Global variable (storing things in memory)
-export const genomes = {} // { hg19: {...}, ... }
+export let genomes // { hg19: {...}, ... }, legacy, should use closure
+export function setGenomes(_genomes) {
+	genomes = _genomes
+}
+
 const tabix = serverconfig.tabix
 const samtools = serverconfig.samtools
 const bcftools = serverconfig.bcftools
@@ -113,220 +106,61 @@ const hicstraw = serverconfig.hicstraw
     */
 const infoFilter_unannotated = 'Unannotated'
 
-export const app = express()
-app.disable('x-powered-by')
+export function setRoutes(app, genomes, serverconfig) {
+	console.log(109, 'setting routes from old app.js')
+	const basepath = serverconfig.basepath || ''
+	// has to set optional routes before app.get() or app.post()
+	// otherwise next() may not be called for a middleware in the optional routes
+	app.get(basepath + '/cardsjson', handle_cards)
+	app.post(basepath + '/mdsjsonform', handle_mdsjsonform)
+	app.post(basepath + '/tkbedj', bedj_request_closure(genomes))
+	app.post(basepath + '/tkbedgraphdot', bedgraphdot_request_closure(genomes))
+	app.all(basepath + '/tkbam', bam_request_closure(genomes))
+	app.get(basepath + '/gdcbam', gdc_bam_request(genomes))
+	app.get(basepath + '/tkaicheck', aicheck_request_closure(genomes))
+	app.get(basepath + '/blat', blat_request_closure(genomes))
+	app.all(basepath + '/mds3', mds3_request_closure(genomes))
+	app.get(basepath + '/tkbampile', bampile_request)
+	app.post(basepath + '/tkld', handle_tkld(genomes))
+	app.get(basepath + '/tabixheader', handle_tabixheader)
+	app.get(basepath + '/img', handle_img)
+	app.post(basepath + '/svmr', handle_svmr)
+	app.post(basepath + '/study', handle_study)
+	app.post(basepath + '/textfile', handle_textfile)
+	app.post(basepath + '/urltextfile', handle_urltextfile)
+	app.get(basepath + '/junction', junction_request) // legacy, including rnapeg
+	app.post(basepath + '/mdsjunction', mdsjunction_request_closure(genomes))
+	app.post(basepath + '/mdscnv', handle_mdscnv)
+	app.post(basepath + '/mdssvcnv', handle_mdssvcnv)
+	app.post(basepath + '/mdsgenecount', handle_mdsgenecount)
+	app.post(basepath + '/mds2', mds2_load.handle_request(genomes))
+	app.post(basepath + '/mdsexpressionrank', handle_mdsexpressionrank) // expression rank as a browser track
+	app.post(basepath + '/mdsgeneboxplot', mdsgeneboxplot_closure(genomes))
+	app.post(basepath + '/mdsgenevalueonesample', handle_mdsgenevalueonesample)
 
-if (serverconfig.users) {
-	// { user1 : pass1, user2: pass2, ... }
-	app.use(basicAuth({ users: serverconfig.users, challenge: true }))
+	app.post(basepath + '/vcf', handle_vcf) // for old ds/vcf and old junction
+
+	app.get(basepath + '/vcfheader', handle_vcfheader)
+
+	app.post(basepath + '/translategm', handle_translategm)
+
+	app.post(basepath + '/samplematrix', handle_samplematrix)
+	app.get(basepath + '/mdssamplescatterplot', handle_mdssamplescatterplot)
+	app.post(basepath + '/mdssamplesignature', handle_mdssamplesignature)
+	app.post(basepath + '/mdssurvivalplot', handle_mdssurvivalplot(genomes))
+	app.post(basepath + '/fimo', fimo.handle_closure(genomes))
+	//app.all(basepath + '/termdb', termdb.handle_request_closure(genomes))
+	app.all(basepath + '/termdb-barsql', termdbbarsql.handle_request_closure(genomes))
+	app.post(basepath + '/singlecell', singlecell.handle_singlecell_closure(genomes))
+	app.post(basepath + '/massSession', massSession.save)
+	app.get(basepath + '/massSession', massSession.get)
+	app.delete(basepath + '/massSession', massSession._delete)
+	app.get(basepath + '/sessionIds', massSession.getSessionIdsByCred)
+	app.get(basepath + '/isoformbycoord', handle_isoformbycoord)
+	app.post(basepath + '/ase', handle_ase)
+	app.post(basepath + '/bamnochr', handle_bamnochr)
+	app.get(basepath + '/ideogram', handle_ideogram)
 }
-
-/* when using webpack, should no longer use __dirname, otherwise cannot find the html files!
-app.use(express.static(__dirname+'/public'))
-*/
-
-export const basepath = serverconfig.basepath || ''
-
-function setHeaders(res) {
-	res.header('Vary', 'Origin')
-	res.header('Access-Control-Allow-Origin', '*')
-	res.header('Access-Control-Allow-Methods', 'GET, POST, DELETE, OPTIONS, HEAD')
-	// embedder sites may use HTTP 2.0 which requires lowercased header key names
-	// must support mixed casing and all lowercased for compatibility
-	res.header(
-		'Access-Control-Allow-Headers',
-		'Origin, X-Requested-With, Content-Type, Accept, Authorization' +
-			', origin, x-requested-with, content-type, accept, authorization' +
-			', X-Auth-Token, X-Ds-Access-Token, X-SjPPDs-Sessionid' +
-			', x-auth-token, x-ds-access-token, x-sjppds-sessionid'
-	)
-}
-
-if (!serverconfig.backend_only) {
-	const staticDir = express.static(path.join(process.cwd(), './public'), { setHeaders })
-	app.use(staticDir)
-}
-
-app.use(compression())
-
-app.use((req, res, next) => {
-	if (req.method.toUpperCase() == 'POST') {
-		// assume all post requests have json-encoded content
-		// TODO: change all client-side fetch(new Request(...)) to use dofetch*() to preset the content-type
-		req.headers['content-type'] = 'application/json'
-	}
-
-	// detect URL parameter values with matching JSON start-stop encoding characters
-	try {
-		const encoding = req.query.encoding
-		urlJsonDecode(req.query)
-	} catch (e) {
-		res.send({ error: e })
-		return
-	}
-	next()
-})
-
-function isNumeric(d) {
-	return !isNaN(parseFloat(d)) && isFinite(d) && d !== ''
-}
-
-app.use(cookieParser())
-app.use(bodyParser.json({ limit: '5mb' }))
-app.use(bodyParser.text({ limit: '5mb' }))
-app.use(bodyParser.urlencoded({ extended: true }))
-
-app.use((req, res, next) => {
-	if (req.method.toUpperCase() == 'POST' && req.body && req.headers['content-type'] != 'application/json') {
-		res.send({ error: `invalid HTTP request.header['content-type'], must be 'application/json'` })
-		return
-	}
-	if (req.headers['content-type'] == 'application/json') {
-		if (!req.query) req.query = {}
-		// TODO: in the future, may have to combine req.query + req.params + req.body
-		// if using req.params based on expressjs server route /:paramName interpolation
-		Object.assign(req.query, req.body)
-	}
-
-	// log the request before adding protected info
-	log(req)
-
-	/*
-	!!! put this code after logging the request, so these protected info are not logged !!!
-	!! more or less quick fix !!
-	in gdc environment, this will pass sessionid from cookie to req.query
-	to be added to request header where it's querying gdc api
-	by doing this, route code is worry-free and no need to pass "req{}" to gdc purpose-specific code doing the API calls
-	these *protected* contents are not used in non-gdc code
-	*/
-	req.query.__protected__ = {}
-	if (req.cookies?.sessionid) {
-		req.query.__protected__.sessionid = req.cookies.sessionid
-	}
-	Object.freeze(req.query.__protected__)
-
-	setHeaders(res)
-	res.header(
-		'Access-Control-Allow-Origin',
-		req.get('origin') || req.get('referrer') || req.protocol + '://' + req.get('host').split(':')[0] || '*'
-	)
-	res.header('Access-Control-Allow-Credentials', true)
-
-	if (req.method == 'GET' && (!req.path.includes('.') || req.path.endsWith('proteinpaint.js'))) {
-		// immutable response before expiration, client must revalidate after max-age;
-		// by convention, any path that has a dot will be treated as
-		// a static file and not handled here with cache-control
-		res.header('Cache-control', `immutable,max-age=${serverconfig.responseMaxAge || 1}`)
-	}
-	next()
-})
-
-app.catch = validator.floodCatch
-app.use(validator.middleware)
-
-// NOTE: The error handling middleware should be used after all other app.use() middlewares
-// error handling middleware has four arguments
-app.use((error, req, res, next) => {
-	log(req)
-	if (error) {
-		if (error && error.type == 'entity.parse.failed') {
-			res.send({ error: 'invalid request body, must be a valid JSON-encoded object' })
-		} else {
-			res.send({ error })
-		}
-		return
-	}
-})
-
-if (serverconfig.jwt) {
-	console.log('JWT is activated')
-	app.use((req, res, next) => {
-		let j = {}
-		if (req.body && req.method == 'POST') {
-			// a preceding middleware assumes all POST contents are json-encoded and processed by bodyParser()
-			j = req.body
-		}
-		const jwt = j.jwt
-			? j.jwt
-			: req.headers && req.headers.authorization && req.headers.authorization.startsWith('Bearer ')
-			? req.headers.authorization.split(' ')[1]
-			: null
-		if (!jwt) return res.send({ error: 'json web token missing' })
-
-		jsonwebtoken.verify(jwt, serverconfig.jwt.secret, (err, decode) => {
-			if (err) return res.send({ error: 'Invalid token' })
-
-			// FIXME do not hardcode required attribute, replace with a list
-			if (!decode[serverconfig.jwt.permissioncheck]) return res.send({ error: 'Not authorized' })
-
-			next()
-		})
-	})
-}
-
-// has to set optional routes before app.get() or app.post()
-// otherwise next() may not be called for a middleware in the optional routes
-setOptionalRoutes()
-authApi.maySetAuthRoutes(app, basepath, serverconfig)
-app.get(basepath + '/cardsjson', handle_cards)
-app.post(basepath + '/mdsjsonform', handle_mdsjsonform)
-app.get(basepath + '/genomes', handle_genomes)
-app.get(basepath + '/getDataset', handle_getDataset)
-app.all(basepath + '/ntseq', handle_ntseq)
-app.post(basepath + '/pdomain', handle_pdomain)
-app.post(basepath + '/tkbedj', bedj_request_closure(genomes))
-app.post(basepath + '/tkbedgraphdot', bedgraphdot_request_closure(genomes))
-app.all(basepath + '/tkbam', bam_request_closure(genomes))
-app.get(basepath + '/gdcbam', gdc_bam_request(genomes))
-app.get(basepath + '/tkaicheck', aicheck_request_closure(genomes))
-app.get(basepath + '/blat', blat_request_closure(genomes))
-app.all(basepath + '/mds3', mds3_request_closure(genomes))
-app.get(basepath + '/tkbampile', bampile_request)
-app.post(basepath + '/dsdata', handle_dsdata) // old official ds, replace by mds
-app.post(basepath + '/tkbigwig', handle_tkbigwig)
-app.post(basepath + '/tkld', handle_tkld(genomes))
-app.get(basepath + '/tabixheader', handle_tabixheader)
-app.all(basepath + '/snp', handle_snp)
-app.post(basepath + '/isoformlst', handle_isoformlst)
-app.post(basepath + '/dbdata', handle_dbdata)
-app.get(basepath + '/img', handle_img)
-app.post(basepath + '/svmr', handle_svmr)
-app.post(basepath + '/study', handle_study)
-app.post(basepath + '/textfile', handle_textfile)
-app.post(basepath + '/urltextfile', handle_urltextfile)
-app.get(basepath + '/junction', junction_request) // legacy, including rnapeg
-app.post(basepath + '/mdsjunction', mdsjunction_request_closure(genomes))
-app.post(basepath + '/mdscnv', handle_mdscnv)
-app.post(basepath + '/mdssvcnv', handle_mdssvcnv)
-app.post(basepath + '/mdsgenecount', handle_mdsgenecount)
-app.post(basepath + '/mds2', mds2_load.handle_request(genomes))
-app.post(basepath + '/mdsexpressionrank', handle_mdsexpressionrank) // expression rank as a browser track
-app.post(basepath + '/mdsgeneboxplot', mdsgeneboxplot_closure(genomes))
-app.post(basepath + '/mdsgenevalueonesample', handle_mdsgenevalueonesample)
-
-app.post(basepath + '/vcf', handle_vcf) // for old ds/vcf and old junction
-
-app.get(basepath + '/vcfheader', handle_vcfheader)
-
-app.post(basepath + '/translategm', handle_translategm)
-
-app.post(basepath + '/samplematrix', handle_samplematrix)
-app.get(basepath + '/mdssamplescatterplot', handle_mdssamplescatterplot)
-app.post(basepath + '/mdssamplesignature', handle_mdssamplesignature)
-app.post(basepath + '/mdssurvivalplot', handle_mdssurvivalplot(genomes))
-app.post(basepath + '/fimo', fimo.handle_closure(genomes))
-app.all(basepath + '/termdb', termdb.handle_request_closure(genomes))
-app.all(basepath + '/termdb-barsql', termdbbarsql.handle_request_closure(genomes))
-app.post(basepath + '/singlecell', singlecell.handle_singlecell_closure(genomes))
-app.post(basepath + '/massSession', massSession.save)
-app.get(basepath + '/massSession', massSession.get)
-app.delete(basepath + '/massSession', massSession._delete)
-app.get(basepath + '/sessionIds', massSession.getSessionIdsByCred)
-app.get(basepath + '/isoformbycoord', handle_isoformbycoord)
-app.post(basepath + '/ase', handle_ase)
-app.post(basepath + '/bamnochr', handle_bamnochr)
-app.get(basepath + '/gene2canonicalisoform', handle_gene2canonicalisoform)
-app.get(basepath + '/ideogram', handle_ideogram)
 
 /****
 	- validate and start the server
@@ -363,79 +197,6 @@ function log(req) {
 		req.header('x-forwarded-for') || req.connection.remoteAddress,
 		JSON.stringify(j).replace(/\\"/g, '"')
 	)
-}
-
-export async function startServer() {
-	// uses the global express "app"
-	try {
-		if (serverconfig.preListenScript) {
-			const { cmd, args } = serverconfig.preListenScript
-			const ps = child_process.spawnSync(cmd, args, { encoding: 'utf-8' })
-			if (ps.stderr.trim()) throw ps.stderr.trim()
-			console.log(ps.stdout)
-		}
-
-		// serverconfig.appEnable is an array of strings that are
-		// valid arguments to http://expressjs.com/en/4x/api.html#app.enabled
-		// For example, when the server sits behind a trusted reverse proxy,
-		// "appEnable": ["trust proxy"]
-		if (serverconfig.appEnable) serverconfig.appEnable.forEach(d => app.enable(d))
-
-		const port = serverconfig.port
-		// !!! DO NOT CHANGE THE FOLLOWING MESSAGE !!!
-		// a serverconfig.preListenScript may rely on detecting this exact post-listen() message
-		const message = `STANDBY AT PORT ${port}`
-		if (serverconfig.ssl) {
-			const options = {
-				key: fs.readFileSync(serverconfig.ssl.key),
-				cert: fs.readFileSync(serverconfig.ssl.cert)
-			}
-			const server = await https.createServer(options, app)
-			// second optional argument is host, formatted so that req.ip will be ipv4
-			server.listen(port, '0.0.0.0', () => {
-				console.log(`HTTPS ${message}`)
-			})
-			return server
-		} else {
-			const server = await http.createServer(app)
-			// second optional argument is host, formatted so that req.ip will be ipv4
-			server.listen(port, '0.0.0.0', () => {
-				if (process.send) {
-					process.send('ready')
-				}
-				console.log(message)
-			})
-			return server
-		}
-	} catch (e) {
-		throw e
-	}
-}
-
-function setOptionalRoutes() {
-	// routeSetters is an array of "filepath/name.js"
-	if (!serverconfig.routeSetters) return
-	for (const fname of serverconfig.routeSetters) {
-		if (fname.endsWith('.js')) {
-			const setRoutes = __non_webpack_require__(fname)
-			setRoutes(app, basepath)
-		}
-	}
-}
-
-function handle_gene2canonicalisoform(req, res) {
-	try {
-		if (!req.query.gene) throw '.gene missing'
-		const genome = genomes[req.query.genome]
-		if (!genome) throw 'unknown genome'
-		if (!genome.genedb.get_gene2canonicalisoform) throw 'gene2canonicalisoform not supported on this genome'
-		const data = genome.genedb.get_gene2canonicalisoform.get(req.query.gene)
-		// data = { isoform: str }
-		res.send(data)
-	} catch (e) {
-		res.send({ error: e.message || e })
-		if (e.stack) console.log(e.stack)
-	}
 }
 
 async function handle_cards(req, res) {
@@ -530,420 +291,6 @@ async function handle_tabixheader(req, res) {
 	}
 }
 
-async function handle_genomes(req, res) {
-	try {
-		await fs.promises.stat(serverconfig.tpmasterdir)
-	} catch (e) {
-		/* dir is inaccessible
-		return error message as the service is out
-		*/
-		// default error message
-		let message = 'Error with TP directory (' + e.code + ')'
-		const m = serverconfig.maintenance || {}
-		// may override with a non-empty maintenance message
-		if ('start' in m && 'stop' in m && m.tpMessage) {
-			// use unix timestamps to simplify comparison
-			const start = +new Date(m.start)
-			const stop = +new Date(m.stop)
-			const currTime = +new Date()
-			if (start <= currTime && currTime <= stop) {
-				message = m.tpMessage
-			}
-		}
-		res.send({ error: message })
-		return
-	}
-
-	const hash = {}
-	if (req.query && req.query.genome) {
-		hash[req.query.genome] = clientcopy_genome(req.query.genome)
-	} else {
-		for (const genomename in genomes) {
-			hash[genomename] = clientcopy_genome(genomename)
-		}
-	}
-	let hasblat = false
-	for (const n in genomes) {
-		if (genomes[n].blat) hasblat = true
-	}
-	res.send({
-		genomes: hash,
-		debugmode: serverconfig.debugmode,
-		headermessage: serverconfig.headermessage,
-		base_zindex: serverconfig.base_zindex,
-		codedate: versionInfo.codedate,
-		launchdate: versionInfo.launchdate,
-		hasblat,
-		features,
-		dsAuth: authApi.getDsAuth(req),
-		commonOverrides: serverconfig.commonOverrides,
-		targetPortal: serverconfig.targetPortal, //sending target portal to the client
-		cardsPath: serverconfig.cards?.path
-	})
-}
-
-function clientcopy_genome(genomename) {
-	const g = genomes[genomename]
-	const g2 = {
-		species: g.species,
-		name: genomename,
-		hasSNP: g.snp ? true : false,
-		hasIdeogram: g.genedb.hasIdeogram,
-		fimo_motif: g.fimo_motif ? true : false,
-		blat: g.blat ? true : false,
-		geneset: g.geneset,
-		defaultcoord: g.defaultcoord,
-		isdefault: g.isdefault,
-		majorchr: g.majorchr,
-		majorchrorder: g.majorchrorder,
-		minorchr: g.minorchr,
-		tracks: g.tracks,
-		hicenzymefragment: g.hicenzymefragment,
-		datasets: {}
-	}
-
-	if (g.termdbs) {
-		g2.termdbs = {}
-		for (const k in g.termdbs) {
-			g2.termdbs[k] = { label: g.termdbs[k].label }
-		}
-	}
-
-	for (const dsname in g.datasets) {
-		const ds = g.datasets[dsname]
-
-		if (ds.isMds3) {
-			// only send most basic info about the dataset, enough for e.g. a button to launch this dataset
-			// client will request detailed info when using this dataset
-			g2.datasets[ds.label] = {
-				isMds3: true,
-				noHandleOnClient: ds.noHandleOnClient,
-				label: ds.label
-			}
-			continue
-		}
-
-		if (ds.isMds) {
-			g2.datasets[ds.label] = {
-				isMds: true,
-				mdsIsUninitiated: true,
-				noHandleOnClient: ds.noHandleOnClient,
-				label: ds.label
-			}
-			continue
-			/*
-			const _ds = mds_clientcopy(ds)
-			if (_ds) {
-				g2.datasets[ds.label] = _ds
-			}
-			continue
-			*/
-		}
-
-		// old official ds; to be replaced by mds3
-		g2.datasets[ds.label] = {
-			isofficial: true,
-			legacyDsIsUninitiated: true, // so client only gets copy_legacyDataset once
-			noHandleOnClient: ds.noHandleOnClient,
-			label: ds.label
-		}
-	}
-
-	if (g.hicdomain) {
-		g2.hicdomain = { groups: {} }
-		for (const s1 in g.hicdomain.groups) {
-			const tt = g.hicdomain.groups[s1]
-			g2.hicdomain.groups[s1] = {
-				name: tt.name,
-				reference: tt.reference,
-				sets: {}
-			}
-			for (const s2 in tt.sets) {
-				g2.hicdomain.groups[s1].sets[s2] = {
-					name: tt.sets[s2].name,
-					longname: tt.sets[s2].longname
-				}
-			}
-		}
-	}
-	return g2
-}
-
-function copy_legacyDataset(ds) {
-	const ds2 = {
-		noHandleOnClient: ds.noHandleOnClient,
-		sampleselectable: ds.sampleselectable,
-		label: ds.label,
-		dsinfo: ds.dsinfo,
-		stratify: ds.stratify,
-		cohort: ds.cohort,
-		vcfinfofilter: ds.vcfinfofilter,
-		info2table: ds.info2table,
-		info2singletable: ds.info2singletable,
-		url4variant: ds.url4variant,
-		itemlabelname: ds.itemlabelname
-	}
-
-	if (ds.snvindel_attributes) {
-		ds2.snvindel_attributes = []
-		for (const at of ds.snvindel_attributes) {
-			const rep = {}
-			for (const k in at) {
-				if (k == 'lst') {
-					rep.lst = []
-					for (const e of at.lst) {
-						const rep2 = {}
-						for (const k2 in e) rep2[k2] = e[k2]
-						rep.lst.push(rep2)
-					}
-				} else {
-					rep[k] = at[k]
-				}
-			}
-			ds2.snvindel_attributes.push(rep)
-		}
-	}
-	if (ds.snvindel_legend) {
-		ds2.snvindel_legend = ds.snvindel_legend
-	}
-	const vcfinfo = {}
-	let hasvcf = false
-	for (const q of ds.queries) {
-		if (q.vcf) {
-			hasvcf = true
-			vcfinfo[q.vcf.vcfid] = q.vcf
-		}
-	}
-	if (hasvcf) {
-		ds2.id2vcf = vcfinfo
-	}
-	return ds2
-}
-
-function handle_getDataset(req, res) {
-	/*
-	q.genome=str, case-sensitive match with genome name
-	q.dsname=str, case-insensitive match with ds key (e.g. pediatric for Pediatric) 
-
-	allow case-insensitive match with dsname
-	*/
-	try {
-		const genome = genomes[req.query.genome]
-		if (!genome) throw 'unknown genome'
-		if (!genome.datasets) throw 'genomeobj.datasets{} missing'
-		let ds
-		for (const k in genome.datasets) {
-			if (k.toLowerCase() == req.query.dsname.toLowerCase()) {
-				ds = genome.datasets[k]
-				break
-			}
-		}
-		if (!ds) throw 'invalid dsname'
-		if (ds.isMds3) {
-			return res.send({ ds: mds3_init.client_copy(ds, null, null, app, basepath) })
-		}
-		if (ds.isMds) {
-			return res.send({ ds: mds_clientcopy(ds) })
-		}
-		return res.send({ ds: copy_legacyDataset(ds) }) // to be replaced by mds3
-	} catch (e) {
-		res.send({ error: e.message || e })
-	}
-}
-
-function mds_clientcopy(ds) {
-	// make client-side copy of a mds
-
-	const ds2 = {
-		isMds: true,
-		noHandleOnClient: ds.noHandleOnClient,
-		label: ds.label,
-		version: ds.version,
-		annotationsampleset2matrix: ds.annotationsampleset2matrix,
-		mutationAttribute: ds.mutationAttribute,
-		locusAttribute: ds.locusAttribute,
-		alleleAttribute: ds.alleleAttribute,
-		// these are quick fixes and should be deleted later
-		hide_genotypedownload: ds.hide_genotypedownload,
-		hide_phewas: ds.hide_phewas,
-		sample2bam: ds.sample2bam
-	}
-
-	if (ds.queries) {
-		ds2.queries = {}
-	}
-
-	if (ds.track) {
-		ds2.track = mds2_init.client_copy(ds)
-	}
-
-	if (ds.singlesamplemutationjson) {
-		ds2.singlesamplemutationjson = 1
-	}
-	if (ds.gene2mutcount) {
-		ds2.gene2mutcount = true
-		ds2.mutCountType = ds.gene2mutcount.mutationTypes
-	}
-	if (ds.assayAvailability) {
-		ds2.assayAvailability = 1
-	}
-
-	if (ds.cohort && ds.cohort.sampleAttribute) {
-		// attr may be hidden from client
-		const toclient = {}
-		for (const k in ds.cohort.sampleAttribute.attributes) {
-			const a = ds.cohort.sampleAttribute.attributes[k]
-			if (!a.clientnoshow) toclient[k] = a
-		}
-		ds2.sampleAttribute = { attributes: toclient }
-	}
-
-	if (ds.cohort) {
-		if (ds.cohort.termdb) {
-			// let client know the existance, do not reveal details unless needed
-			ds2.termdb = {
-				selectCohort: ds.cohort.termdb.selectCohort
-			}
-		}
-
-		if (ds.cohort.attributes && ds.cohort.attributes.defaulthidden) {
-			/*
-            .attributes.lst[] are not released to client
-            default hidden attributes from sample annotation, tell client
-            */
-			ds2.cohortHiddenAttr = ds.cohort.attributes.defaulthidden
-		}
-
-		if (ds.cohort.survivalplot) {
-			ds2.survivalplot = {
-				samplegroupattrlst: ds.cohort.survivalplot.samplegroupattrlst,
-				plots: []
-			}
-			for (const k in ds.cohort.survivalplot.plots) {
-				const p = ds.cohort.survivalplot.plots[k]
-				ds2.survivalplot.plots.push({
-					key: k,
-					name: p.name,
-					timelabel: p.timelabel
-				})
-			}
-		}
-
-		if (ds.cohort.mutation_signature) {
-			const sets = {}
-			for (const k in ds.cohort.mutation_signature.sets) {
-				const s = ds.cohort.mutation_signature.sets[k]
-				sets[k] = {
-					name: s.name,
-					signatures: s.signatures
-				}
-			}
-			ds2.mutation_signature = { sets: sets }
-		}
-	}
-
-	for (const k in ds.queries) {
-		const q = ds.queries[k]
-
-		const clientquery = {
-			// revealed to client
-			name: q.name,
-			hideforthemoment: q.hideforthemoment // hide track not ready to show on client
-		}
-
-		if (q.istrack) {
-			clientquery.istrack = true
-			clientquery.type = q.type
-			clientquery.isfull = q.isfull
-			// track attributes, some are common, many are track type-specific
-			if (q.nochr != undefined) {
-				clientquery.nochr = q.nochr
-			}
-			if (q.infoFilter) {
-				clientquery.infoFilter = q.infoFilter
-			}
-			// junction attributes
-			if (q.readcountCutoff) {
-				clientquery.readcountCutoff = q.readcountCutoff
-			}
-			// cnv attributes
-			if (q.valueLabel) {
-				clientquery.valueLabel = q.valueLabel
-			}
-			if (q.valueCutoff) {
-				clientquery.valueCutoff = q.valueCutoff
-			}
-			if (q.bplengthUpperLimit) {
-				clientquery.bplengthUpperLimit = q.bplengthUpperLimit
-			}
-			// loh attributes
-			if (q.segmeanValueCutoff) {
-				clientquery.segmeanValueCutoff = q.segmeanValueCutoff
-			}
-			if (q.lohLengthUpperLimit) {
-				clientquery.lohLengthUpperLimit = q.lohLengthUpperLimit
-			}
-
-			if (q.type == common.tkt.mdssvcnv) {
-				if (q.groupsamplebyattr) {
-					clientquery.groupsamplebyattr = q.groupsamplebyattr
-				}
-
-				// flags
-				clientquery.multihidelabel_fusion = q.multihidelabel_fusion
-				clientquery.multihidelabel_sv = q.multihidelabel_sv
-				clientquery.multihidelabel_vcf = q.multihidelabel_vcf
-				clientquery.showfullmode = q.showfullmode
-				clientquery.legend_vorigin = q.legend_vorigin
-				clientquery.no_loh = q.no_loh // quick dirty fix
-
-				if (q.expressionrank_querykey) {
-					// for checking expression rank
-					const e = ds.queries[q.expressionrank_querykey]
-					clientquery.checkexpressionrank = {
-						querykey: q.expressionrank_querykey,
-						datatype: e.datatype
-					}
-					if (e.boxplotbysamplegroup && e.boxplotbysamplegroup.additionals) {
-						// quick fix!!
-						// array element 0 is boxplotbysamplegroup.attributes
-						// rest of array, one ele for each of .additionals
-						const lst = []
-						if (e.boxplotbysamplegroup.attributes)
-							lst.push(e.boxplotbysamplegroup.attributes.map(i => i.label).join(', '))
-						for (const i of e.boxplotbysamplegroup.additionals) lst.push(i.label)
-						clientquery.checkexpressionrank.boxplotgroupers = lst
-					}
-				}
-				if (q.vcf_querykey) {
-					clientquery.checkvcf = {
-						querykey: q.vcf_querykey,
-						info: ds.queries[q.vcf_querykey].info,
-						format: {}
-					}
-					for (const tk of ds.queries[q.vcf_querykey].tracks) {
-						if (tk.format) {
-							for (const k in tk.format) {
-								clientquery.checkvcf.format[k] = tk.format[k]
-							}
-						}
-					}
-				}
-			}
-		} else if (q.isgenenumeric) {
-			clientquery.isgenenumeric = true
-			clientquery.datatype = q.datatype
-			clientquery.no_ase = q.no_ase
-		} else {
-			// this query is not to be revealed to client
-			continue
-		}
-
-		ds2.queries[k] = clientquery
-	}
-	return ds2
-}
-
 async function handle_img(req, res) {
 	const [e, file, isurl] = utils.fileurl(req) // utils.fileurl({ query: { file: req.query.file } })
 	try {
@@ -955,313 +302,6 @@ async function handle_img(req, res) {
 		})
 	} catch (e) {
 		res.send({ error: e.message || e })
-	}
-}
-
-async function handle_ntseq(req, res) {
-	try {
-		if (!req.query.coord) throw 'coord missing'
-		const g = genomes[req.query.genome]
-		if (!g) throw 'invalid genome'
-		if (!g.genomefile) throw 'no sequence file available'
-		const seq = await utils.get_fasta(g, req.query.coord)
-		res.send({
-			seq: seq.split('\n').slice(1).join('')
-		})
-	} catch (e) {
-		res.send({ error: e.message || e })
-		if (e.stack) console.log(e.stack)
-	}
-}
-
-function handle_pdomain(req, res) {
-	try {
-		const gn = req.query.genome
-		if (!gn) throw 'no genome'
-		const g = genomes[gn]
-		if (!g) throw 'invalid genome ' + gn
-		if (!g.proteindomain) {
-			// no error
-			return res.send({ lst: [] })
-		}
-		if (!Array.isArray(req.query.isoforms)) throw 'isoforms[] missing'
-		const lst = []
-		for (const isoform of req.query.isoforms) {
-			if (g.genomicNameRegexp.test(isoform)) continue
-			const tmp = g.proteindomain.getbyisoform.all(isoform)
-			// FIXME returned {data} is text not json
-			lst.push({
-				name: isoform,
-				pdomains: tmp.map(i => {
-					const j = JSON.parse(i.data)
-					j.refseq = isoform
-					return j
-				})
-			})
-		}
-		res.send({ lst })
-	} catch (e) {
-		res.send({ error: e.message || e })
-		if (e.stack) console.log(e.stack)
-	}
-}
-
-async function handle_snp(req, res) {
-	// TODO move to routes
-	try {
-		const n = req.query.genome
-		if (!n) throw 'no genome'
-		res.send({ results: await searchSNP(req.query, genomes[n]) })
-	} catch (e) {
-		if (e.stack) console.log(e.stack)
-		return res.send({ error: e.message || e })
-	}
-}
-
-/*
-.byCoord
-	if true, query bigbed file by coordinate
-.byName
-	if true, query bigbed file by rs id
-.genome
-	name of genome
-.chr
-	used for byCoord
-.ranges[ {start, stop} ]
-	used for byCoord
-.alleleLst[ str ]
-	used for byCoord, if provided will only return snps with matching alleles
-.lst[ str ]
-	used for byName
-*/
-export async function searchSNP(q, genome) {
-	if (!genome) throw 'invalid genome'
-	if (!genome.snp) throw 'snp is not configured for this genome'
-	const hits = []
-	if (q.byCoord) {
-		// query dbSNP bigbed file by coordinate
-		// input query coordinates need to be 0-based
-		// output snp coordinates are 0-based
-		if (genome.genomicNameRegexp.test(q.chr)) throw 'invalid chr name'
-		if (!Array.isArray(q.ranges)) throw 'ranges not an array'
-		for (const r of q.ranges) {
-			// require start/stop of a range to be non-neg integers, as a measure against attack
-			if (!Number.isInteger(r.start) || !Number.isInteger(r.stop) || r.start < 0 || r.stop < r.start)
-				throw 'invalid start/stop'
-			if (r.stop - r.start >= 100) {
-				// quick fix!
-				// as this function only works as spot checking snps
-				// guard against big range and avoid retrieving snps from whole chromosome that will overwhelm server
-				throw 'range too big'
-			}
-			const snps = await utils.query_bigbed_by_coord(genome.snp.bigbedfile, q.chr, r.start, r.stop)
-			for (const snp of snps) {
-				const hit = snp2hit(snp)
-				if (q.alleleLst) {
-					// given alleles must be found in a snp for it to be returned
-					let missing = false
-					for (const i of q.alleleLst) {
-						// only test on non-empty strings
-						if (i && !hit.alleles.includes(i)) {
-							missing = true
-							break
-						}
-					}
-					if (missing) continue
-				}
-				hits.push(hit)
-			}
-		}
-	} else if (q.byName) {
-		if (!Array.isArray(q.lst)) throw '.lst[] missing'
-		for (const n of q.lst) {
-			// query dbSNP bigbed file by rsID
-			// see above for description of output snp fields
-			if (genome.genomicNameRegexp.test(n)) continue
-			const snps = await utils.query_bigbed_by_name(genome.snp.bigbedfile, n)
-			for (const snp of snps) {
-				const hit = snp2hit(snp)
-				hits.push(hit)
-			}
-		}
-	} else {
-		throw 'unknown query method'
-	}
-	return hits
-}
-
-function snp2hit(snp) {
-	// snp must be non-empty string
-	// output snp fields: [0]chrom, [1]chromStart, [2]chromEnd, [3]name, [4]ref, altCount, [6]alts, shiftBases, freqSourceCount, minorAlleleFreq, majorAllele, minorAllele, maxFuncImpact, class, ucscNotes, _dataOffset, _dataLen
-	const fields = snp.split('\t')
-	const ref = fields[4]
-	const alts = fields[6].split(',').filter(Boolean)
-	const observed = ref + '/' + alts.join('/')
-	const hit = {
-		chrom: fields[0],
-		chromStart: Number(fields[1]),
-		chromEnd: Number(fields[2]),
-		name: fields[3],
-		observed: observed,
-		alleles: [ref, ...alts]
-	}
-	return hit
-}
-
-async function handle_dsdata(req, res) {
-	/*
-    poor mechanism, only for old-style official dataset
-
-    to be totally replaced by mds, which can identify queries in a mds by querykeys
-    */
-	try {
-		if (!genomes[req.query.genome]) throw 'invalid genome'
-		if (!req.query.dsname) throw '.dsname missing'
-		const ds = genomes[req.query.genome].datasets[req.query.dsname]
-		if (!ds) throw 'invalid dsname'
-
-		const data = []
-
-		for (const query of ds.queries) {
-			if (req.query.expressiononly && !query.isgeneexpression) {
-				/*
-                expression data only
-                TODO mds should know exactly which data type to query, or which vending button to use
-                */
-				continue
-			}
-			if (req.query.noexpression && query.isgeneexpression) {
-				// skip expression data
-				continue
-			}
-
-			if (query.dsblocktracklst) {
-				/*
-                do not load any tracks here yet
-                TODO should allow loading some/all, when epaint is not there
-                */
-				continue
-			}
-
-			if (query.vcffile) {
-				const d = await handle_dsdata_vcf(query, req)
-				data.push(d)
-				continue
-			}
-
-			if (query.makequery) {
-				const d = handle_dsdata_makequery(ds, query, req)
-				data.push(d)
-				continue
-			}
-
-			throw 'unknow type from one of ds.queries[]'
-		}
-
-		res.send({ data })
-	} catch (e) {
-		if (e.stack) console.log(e.stack)
-		res.send({ error: e.message || e })
-	}
-}
-
-function handle_dsdata_makequery(ds, query, req) {
-	// query from ds.newconn
-
-	if (req.query.isoform) {
-		// quick fix!! deflect attacks from isoform parameter to avoid db query
-		if (genomes[req.query.genome].genomicNameRegexp.test(req.query.isoform)) return
-	}
-
-	const [sqlstr, values] = query.makequery(req.query)
-	if (!sqlstr) {
-		// when not using gm, will not query tables such as expression
-		return
-	}
-	const rows = ds.newconn.prepare(sqlstr).all(values)
-	let lst
-	if (query.tidy) {
-		lst = rows.map(i => query.tidy(i))
-	} else {
-		lst = rows
-	}
-	const result = {}
-	if (query.isgeneexpression) {
-		result.lst = lst
-		result.isgeneexpression = true
-		result.config = query.config
-
-		/*
-        	loading of junction track as a dependent of epaint
-        	attach junction track info in this result, for making the junction button in epaint
-        	await user to click that button
-
-        	replace-by-mds
-
-        	*/
-
-		for (const q2 of ds.queries) {
-			if (!q2.dsblocktracklst) continue
-			for (const tk of q2.dsblocktracklst) {
-				if (tk.type == common.tkt.junction) {
-					result.config.dsjunctiontk = tk
-				}
-			}
-		}
-	} else {
-		result.lst = lst
-	}
-	return result
-}
-
-function handle_dsdata_vcf(query, req) {
-	const par = [
-		path.join(serverconfig.tpmasterdir, query.vcffile),
-		(query.vcf.nochr ? req.query.range.chr.replace('chr', '') : req.query.range.chr) +
-			':' +
-			req.query.range.start +
-			'-' +
-			req.query.range.stop
-	]
-	return new Promise((resolve, reject) => {
-		const ps = spawn(tabix, par)
-		const out = [],
-			out2 = []
-		ps.stdout.on('data', i => out.push(i))
-		ps.stderr.on('data', i => out2.push(i))
-		ps.on('close', code => {
-			const e = out2.join('').trim()
-			if (e != '') reject('error querying vcf file')
-			const tmp = out.join('').trim()
-			resolve({
-				lines: tmp == '' ? [] : tmp.split('\n'),
-				vcfid: query.vcf.vcfid
-			})
-		})
-	})
-}
-
-function handle_isoformlst(req, res) {
-	try {
-		const g = genomes[req.query.genome]
-		if (!g) throw 'invalid genome'
-		if (!Array.isArray(req.query.lst)) throw '.lst missing'
-		const lst = []
-		for (const isoform of req.query.lst) {
-			if (g.genomicNameRegexp.test(isoform)) continue
-			const tmp = g.genedb.getjsonbyisoform.all(isoform)
-			lst.push(
-				tmp.map(i => {
-					const j = JSON.parse(i.genemodel)
-					if (i.isdefault) j.isdefault = true
-					return j
-				})
-			)
-		}
-		res.send({ lst })
-	} catch (e) {
-		res.send({ error: e.message || e })
-		if (e.stack) console.log(e.stack)
 	}
 }
 
@@ -6535,1811 +5575,6 @@ function parse_textfilewithheader(text) {
 }
 
 /***************************   end of __util   **/
-
-/* when starting pp server process with "npm start" or "node server.js"
-pp_init() runs first to load all genomes supported on this server,
-and load all datasets supported in each genome
-as encoded in file "serverconfig.json"
-at the end it will 
-*/
-export async function pp_init(serverconfig) {
-	// verify if tp directory is readable
-	// ppr has this situation where its tp/ is from a nfs mount and can go down...
-	try {
-		await fs.promises.stat(serverconfig.tpmasterdir)
-	} catch (e) {
-		/* dir is inaccessible for some reason */
-		const message = 'Error with ' + serverconfig.tpmasterdir + ': ' + e.code
-		if (process.argv[2] == 'validate') {
-			throw message
-		} else {
-			// allow the server process to boot
-			// we want the node server to keep running so it can inform user with some meaningful msg rather than http error
-			console.log('\n!!! ' + message + '\n')
-			return
-		}
-	}
-
-	checkDependenciesAndVersions()
-
-	// date updated
-
-	if (serverconfig.base_zindex != undefined) {
-		const v = Number.parseInt(serverconfig.base_zindex)
-		if (Number.isNaN(v) || v <= 0) throw 'base_zindex must be positive integer'
-		serverconfig.base_zindex = v
-	}
-	if (serverconfig.jwt) {
-		if (!serverconfig.jwt.secret) throw 'jwt.secret missing'
-		if (!serverconfig.jwt.permissioncheck) throw 'jwt.permissioncheck missing'
-	}
-	if (!serverconfig.tpmasterdir) throw '.tpmasterdir missing'
-	if (!serverconfig.cachedir) throw '.cachedir missing'
-
-	// create sub directories under cachedir, and register path in serverconfig
-	// to ensure temp files saved in previous server session are accessible in current session
-	// must use consistent dir name but not random dir name that changes from last server boot
-	serverconfig.cachedir_massSession = await mayCreateSubdirInCache('massSession')
-
-	//DELETE THIS after process for deleting mass session files moved into production
-	serverconfig.cachedir_massSessionTrash = await mayCreateSubdirInCache('massSessionTrash')
-
-	serverconfig.cache_snpgt = {
-		dir: await mayCreateSubdirInCache('snpgt'),
-		fileNameRegexp: /[^\w]/, // client-provided cache file name matching with this are denied
-		sampleColumn: 6 // in cache file, sample column starts from 7th column
-	}
-	serverconfig.cachedir_bam = await mayCreateSubdirInCache('bam')
-	serverconfig.cachedir_genome = await mayCreateSubdirInCache('genome')
-	serverconfig.cachedir_ssid = await mayCreateSubdirInCache('ssid')
-
-	// NOTE: required or imported code files are only loaded once by Nodejs
-	// and variables are static so that changes to common key-values will affect all
-	// server-side code that import common.js
-	if (serverconfig.commonOverrides) {
-		common.applyOverrides(serverconfig.commonOverrides)
-	}
-
-	if (!serverconfig.genomes) throw '.genomes[] missing'
-	if (!Array.isArray(serverconfig.genomes)) throw '.genomes[] not array'
-
-	/*
-	for genomes declared in serverconfig for this pp instance,
-	load its built in genome javascript file to an in-mem object
-	and apply overrides from serverconfig to modify this object
-	keep the object in genomes{}
-	*/
-	for (const g of serverconfig.genomes) {
-		if (!g.name) throw '.name missing from a genome: ' + JSON.stringify(g)
-		if (!g.file) throw '.file missing from genome ' + g.name
-
-		/*
-			When using a Docker container, the mounted app directory
-			may have an optional genome directory, which if present
-			will be symlinked to the app directory and potentially override any
-			similarly named genome js file that are part of the standard
-			Proteinpaint packaged files[] 
-		*/
-		const overrideFile = path.join(process.cwd(), g.file)
-
-		/* g is the object from serverconfig.json, for instance-specific customizations of this genome
-		g2 is the standard-issue obj loaded from the js file
-		settings in g will modify g2
-		g2 is registered in the global "genomes"
-		*/
-		const g2module = __non_webpack_require__(fs.existsSync(overrideFile) ? overrideFile : g.file)
-		const g2 = g2module.default || g2module
-		genomes[g.name] = g2
-
-		if (!g2.genomefile) throw '.genomefile missing from .js file of genome ' + g.name
-		if (g2.genomefile == 'NA') {
-			// not available
-		} else {
-			g2.genomefile = path.join(serverconfig.tpmasterdir, g2.genomefile)
-		}
-
-		// for testing if gene/isoform/chr/snp names have only allowed characters
-		// test to true if name has extra characters which could be attack strings
-		// allow for genome-specific pattern setting, otherwise default is used
-		if (!g2.genomicNameRegexp) g2.genomicNameRegexp = /[^a-zA-Z0-9.:_-]/
-
-		if (!g2.tracks) {
-			g2.tracks = [] // must always have .tracks even if empty
-		}
-		if (g.tracks) {
-			// supplement
-			for (const t of g.tracks) {
-				g2.tracks.push(t)
-			}
-		}
-		if (g.datasets) {
-			g2.rawdslst = g.datasets
-		}
-		if (g.snp) {
-			// replace snp db
-			g2.snp = g.snp
-		}
-		if (g.blat) {
-			if (!g.blat.host) throw '.blat.host missing for ' + g.name
-			if (!g.blat.port) throw '.blat.port missing for ' + g.name
-			// gfServer must start with full path to 2bit file
-			g2.blat = g.blat // enable blat
-		}
-		if (g.nosnp) {
-			// no snp
-			delete g2.snp
-		}
-		if (g.nohicenzyme) {
-			delete g2.hicenzymefragment
-		}
-		if (g.nohicdomain) {
-			delete g2.hicdomain
-		}
-		if (g2.genedb) {
-			if (g.no_gene2canonicalisoform) delete g2.genedb.gene2canonicalisoform
-		}
-
-		if (g.updateAttr) {
-			for (const row of g.updateAttr) {
-				let pointer = g2
-				for (const [i, field] of row.entries()) {
-					// to guard against invalid keys, could be manual errors or updated dataset spec
-					if (!pointer) continue
-
-					if (typeof field == 'object') {
-						// apply the key-value overrides to the object that is pointed to
-						for (const k in field) {
-							pointer[k] = field[k]
-						}
-					} else {
-						if (typeof pointer[field] == 'string') {
-							//terminal
-							if (row[i + 1]) {
-								pointer[field] = row[i + 1]
-							}
-							break
-						}
-
-						// reset the reference to a subnested object
-						pointer = pointer[field]
-					}
-				}
-			}
-		}
-	}
-
-	if (serverconfig.defaultgenome) {
-		if (genomes[serverconfig.defaultgenome]) {
-			genomes[serverconfig.defaultgenome].isdefault = true
-		}
-	}
-
-	for (const genomename in genomes) {
-		/*
-		validate each genome
-		*/
-		const g = genomes[genomename]
-		if (!g.majorchr) throw genomename + ': majorchr missing'
-		if (!g.defaultcoord) throw genomename + ': defaultcoord missing'
-
-		// test samtools and genomefile
-		await utils.get_fasta(g, g.defaultcoord.chr + ':' + g.defaultcoord.start + '-' + (g.defaultcoord.start + 1))
-
-		if (!g.tracks) {
-			g.tracks = []
-		}
-		if (typeof g.majorchr == 'string') {
-			const lst = g.majorchr.trim().split(/[\s\t\n]+/)
-			const hash = {}
-			const chrorder = []
-			for (let i = 0; i < lst.length; i += 2) {
-				const chr = lst[i]
-				const c = Number.parseInt(lst[i + 1])
-				if (Number.isNaN(c)) throw genomename + ' majorchr invalid chr size for ' + chr + ' (' + lst[i + 1] + ')'
-				hash[chr] = c
-				chrorder.push(chr)
-			}
-			g.majorchr = hash
-			g.majorchrorder = chrorder
-		}
-		g.chrlookup = {}
-		// k: uppercase chr, v: {name:str, len:int, major:bool}
-		for (const n in g.majorchr) {
-			g.chrlookup[n.toUpperCase()] = { name: n, len: g.majorchr[n], major: true }
-		}
-		if (g.minorchr) {
-			if (typeof g.minorchr == 'string') {
-				const lst = g.minorchr.trim().split(/[\s\t\n]+/)
-				const hash = {}
-				for (let i = 0; i < lst.length; i += 2) {
-					const c = Number.parseInt(lst[i + 1])
-					if (Number.isNaN(c)) throw genomename + ' minorchr invalid chr size for ' + lst[i] + ' (' + lst[i + 1] + ')'
-					hash[lst[i]] = c
-				}
-				g.minorchr = hash
-			}
-			for (const n in g.minorchr) {
-				g.chrlookup[n.toUpperCase()] = { name: n, len: g.minorchr[n] }
-			}
-		}
-
-		// genedb is optional
-		if (g.genedb) {
-			if (!g.genedb.dbfile) throw genomename + ': .genedb.dbfile missing'
-			// keep reference of the connection (.db) so as to add dataset-specific query statements later
-			try {
-				console.log('Connecting', g.genedb.dbfile)
-				g.genedb.db = utils.connect_db(g.genedb.dbfile)
-			} catch (e) {
-				throw `Cannot connect genedb: ${g.genedb.dbfile}: ${e}`
-			}
-			g.genedb.getnamebynameorisoform = g.genedb.db.prepare('select name from genes where name=? or isoform=?')
-			g.genedb.getnamebyisoform = g.genedb.db.prepare('select distinct name from genes where isoform=?')
-			g.genedb.getjsonbyname = g.genedb.db.prepare('select isdefault,genemodel from genes where name=?')
-			g.genedb.getjsonbyisoform = g.genedb.db.prepare('select isdefault,genemodel from genes where isoform=?')
-			g.genedb.getnameslike = g.genedb.db.prepare('select distinct name from genes where name like ? limit 20')
-
-			/*
-			optional tables in gene db:
-
-			- genealias
-			- gene2coord
-			- ideogram
-			- gene2canonicalisoform
-			- refseq2ensembl
-			- buildDate
-
-			if present, create getter to this table and attach to g.genedb{}
-			*/
-			const tables = listDbTables(g.genedb.db)
-			if (tables.has('genealias')) {
-				g.genedb.getNameByAlias = g.genedb.db.prepare('select name from genealias where alias=?')
-				// quick fix -- convert symbol to ENSG, to be used for gdc api query
-				g.genedb.getAliasByName = g.genedb.db.prepare('select alias from genealias where name=?')
-			}
-			if (tables.has('gene2coord')) {
-				g.genedb.getCoordByGene = g.genedb.db.prepare('select * from gene2coord where name=?')
-			}
-			if (tables.has('ideogram')) {
-				g.genedb.hasIdeogram = true
-				g.genedb.getIdeogramByChr = g.genedb.db.prepare('select * from ideogram where chromosome=?')
-			} else {
-				g.genedb.hasIdeogram = false
-			}
-			if (tables.has('gene2canonicalisoform')) {
-				g.genedb.get_gene2canonicalisoform = g.genedb.db.prepare(
-					'select isoform from gene2canonicalisoform where gene=?'
-				)
-			}
-			if (tables.has('buildDate')) {
-				g.genedb.get_buildDate = g.genedb.db.prepare('select date from buildDate')
-			}
-
-			// this table is only used for gdc dataset
-			g.genedb.hasTable_refseq2ensembl = tables.has('refseq2ensembl')
-
-			g.genedb.sqlTables = [...tables]
-			g.genedb.tableSize = {}
-			for (const table of tables) {
-				if (table == 'buildDate') continue
-				g.genedb.tableSize[table] = g.genedb.db.prepare(`select count(*) as size from ${table}`).get().size
-			}
-		}
-
-		// termdbs{} is optional
-		if (g.termdbs) {
-			for (const key in g.termdbs) {
-				server_init_db_queries(g.termdbs[key], features)
-				console.log(`${key} initiated as ${genomename}-level termdb`)
-			}
-		}
-
-		for (const tk of g.tracks) {
-			if (!tk.__isgene) continue
-			if (!tk.file) throw 'Tabix file missing for gene track: ' + JSON.stringify(tk)
-			try {
-				await utils.validate_tabixfile(path.join(serverconfig.tpmasterdir, tk.file))
-			} catch (e) {
-				throw 'Error with ' + tk.file + ': ' + e
-			}
-		}
-
-		if (g.proteindomain) {
-			if (!g.proteindomain.dbfile) throw genomename + '.proteindomain: missing dbfile for sqlite db'
-			if (!g.proteindomain.statement) throw genomename + '.proteindomain: missing statement for sqlite db'
-			let db
-			try {
-				console.log('Connecting', g.proteindomain.dbfile)
-				db = utils.connect_db(g.proteindomain.dbfile)
-			} catch (e) {
-				throw 'Error with ' + g.proteindomain.dbfile + ': ' + e
-			}
-			g.proteindomain.getbyisoform = db.prepare(g.proteindomain.statement)
-		}
-
-		if (g.snp) {
-			if (!g.snp.bigbedfile) throw genomename + '.snp: missing bigBed file'
-			g.snp.bigbedfile = path.join(serverconfig.tpmasterdir, g.snp.bigbedfile)
-			await utils.file_is_readable(g.snp.bigbedfile)
-		}
-
-		if (g.fimo_motif) {
-			if (!g.fimo_motif.db) throw genomename + '.fimo_motif: db file missing'
-			g.fimo_motif.db = path.join(serverconfig.tpmasterdir, g.fimo_motif.db)
-			if (g.fimo_motif.annotationfile) {
-				const [err, items] = parse_textfilewithheader(
-					fs.readFileSync(path.join(serverconfig.tpmasterdir, g.fimo_motif.annotationfile), { encoding: 'utf8' }).trim()
-				)
-				g.fimo_motif.tf2attr = {}
-				for (const i of items) {
-					g.fimo_motif.tf2attr[i.Model.split('_')[0]] = i
-				}
-			}
-		}
-
-		if (g.hicenzymefragment) {
-			if (!Array.isArray(g.hicenzymefragment)) throw 'hicenzymefragment should be an array'
-			for (const frag of g.hicenzymefragment) {
-				if (!frag.enzyme) throw '.enzyme missing for one element of hicenzymefragment[]'
-				if (!frag.file) throw '.file missing for one element of hicenzymefragment[]'
-				try {
-					await utils.validate_tabixfile(path.join(serverconfig.tpmasterdir, frag.file))
-				} catch (e) {
-					throw 'Error with ' + frag.file + ': ' + e
-				}
-			}
-		}
-
-		if (g.hicdomain) {
-			if (!g.hicdomain.groups) throw '.groups{} missing from hicdomain'
-			for (const groupname in g.hicdomain.groups) {
-				const grp = g.hicdomain.groups[groupname]
-				if (!grp.name) throw '.name missing from hicdomain ' + groupname
-				if (!grp.sets) throw '.set{} missing from hicdomain ' + groupname
-				for (const setname in grp.sets) {
-					const hs = grp.sets[setname]
-					if (!hs.name) throw '.name missing from hicdomain ' + groupname + ' > ' + setname
-					if (!hs.file) throw '.file missing from hicdomain ' + groupname + ' > ' + setname
-					hs.file = path.join(serverconfig.tpmasterdir, hs.file) // replace with full path, keep on server side
-					try {
-						await utils.validate_tabixfile(hs.file)
-					} catch (e) {
-						throw 'Error with ' + hs.file + ': ' + e
-					}
-				}
-			}
-		}
-
-		if (!g.rawdslst) {
-			// allow to have no ds
-			continue
-		}
-
-		/*
-	done everything except dataset
-	*/
-
-		g.datasets = {}
-		for (const d of g.rawdslst) {
-			/*
-		for each raw dataset
-		*/
-			if (d.skip) continue
-			if (!d.name) throw 'a nameless dataset from ' + genomename
-			if (g.datasets[d.name]) throw genomename + ' has duplicating dataset name: ' + d.name
-			if (!d.jsfile) throw 'jsfile not available for dataset ' + d.name + ' of ' + genomename
-
-			/*
-				When using a Docker container, the mounted app directory
-				may have an optional dataset directory, which if present
-				will be symlinked to the app directory and potentially override any
-				similarly named dataset js file that are part of the standard
-				Proteinpaint packaged files[] 
-			*/
-			const overrideFile = path.join(process.cwd(), d.jsfile)
-			const _ds = __non_webpack_require__(fs.existsSync(overrideFile) ? overrideFile : d.jsfile)
-			const ds =
-				typeof _ds == 'function'
-					? _ds(common)
-					: typeof _ds?.default == 'function'
-					? _ds.default(common)
-					: _ds.default || _ds
-
-			// !!! TODO: is this unnecessarily repeated at a later time? !!!
-			server_updateAttr(ds, d)
-			ds.noHandleOnClient = d.noHandleOnClient
-			ds.label = d.name
-			ds.genomename = genomename
-			g.datasets[ds.label] = ds
-
-			if (ds.isMds3) {
-				try {
-					await mds3_init.init(ds, g, d, app, basepath)
-				} catch (e) {
-					if (e.stack) console.log(e.stack)
-					throw 'Error with mds3 dataset ' + ds.label + ': ' + e
-				}
-				continue
-			}
-			if (ds.isMds) {
-				try {
-					await mds_init(ds, g, d)
-				} catch (e) {
-					throw 'Error with mds dataset ' + ds.label + ': ' + e
-				}
-				continue
-			}
-
-			initLegacyDataset(ds, g)
-		}
-
-		deleteSessionFiles()
-
-		delete g.rawdslst
-	}
-}
-
-async function deleteSessionFiles() {
-	//Delete mass session files older than the massSessionDuration in serverconfig or 30 days default
-	const files = await fs.promises.readdir(serverconfig.cachedir_massSession)
-	try {
-		files.forEach(async file => {
-			//Return creation Time
-			const stats = await fs.promises.stat(path.join(serverconfig.cachedir_massSession, file))
-			const sessionCreationDate = stats.birthtime
-
-			//Determine file age against massSessionDuration
-			const today = new Date()
-			const fileDate = new Date(sessionCreationDate)
-			const massSessionDuration = serverconfig.features.massSessionDuration || 30
-			const sessionDaysElapsed = Math.round((today.getTime() - fileDate.getTime()) / (1000 * 3600 * 24))
-			if (sessionDaysElapsed > massSessionDuration) {
-				// Move file to massSessionTrash
-				// Process in place until users get use to it
-				await fs.promises.copyFile(
-					path.join(serverconfig.cachedir_massSession, file),
-					path.join(serverconfig.cachedir_massSessionTrash, file)
-				)
-				// Delete file out of massSession
-				await fs.promises.unlink(path.join(serverconfig.cachedir_massSession, file))
-
-				console.log('File deleted: ', file, sessionCreationDate)
-			}
-		})
-	} catch (e) {
-		throw `Error: ${e}`
-	}
-}
-
-function checkDependenciesAndVersions() {
-	if (serverconfig.features.skip_checkDependenciesAndVersions) {
-		console.log('SKIPPED checkDependenciesAndVersions()')
-		return
-	}
-
-	// test if R has all required libraries
-	const rlibraries = ['jsonlite', 'cmprsk', 'hwde', 'lmtest']
-	for (const lib of rlibraries) {
-		const ps = child_process.spawnSync(
-			serverconfig.Rscript,
-			['-e', `suppressPackageStartupMessages(library("${lib}"))`],
-			{ encoding: 'utf8' }
-		)
-		if (ps.stderr.trim()) throw ps.stderr
-	}
-
-	// samtools and bcftools usually have similar installed versions
-	const htslibMinorVer = 10
-	{
-		const lines = child_process
-			.execSync(serverconfig.samtools + ' --version', { encoding: 'utf8' })
-			.trim()
-			.split('\n')
-		// first line should be "samtools 1.14"
-		const [name, v] = lines[0].split(' ')
-		if (name != 'samtools' || !v) throw 'cannot run "samtools version"'
-		const [major, minor] = v.split('.')
-		if (major != '1') throw 'samtools not 1.*'
-		const i = Number(minor)
-		if (i < htslibMinorVer) throw `samtools not >= 1.${htslibMinorVer}`
-	}
-	{
-		const lines = child_process
-			.execSync(serverconfig.bcftools + ' -v', { encoding: 'utf8' })
-			.trim()
-			.split('\n')
-		// first line should be "bcftools 1.14"
-		const [name, v] = lines[0].split(' ')
-		if (name != 'bcftools' || !v) throw 'cannot run "bcftools version"'
-		const [major, minor] = v.split('.')
-		if (major != '1') throw 'bcftools not 1.*'
-		const i = Number(minor)
-		if (i < htslibMinorVer) throw `bcftools not >= 1.${htslibMinorVer}`
-	}
-}
-
-function initLegacyDataset(ds, genome) {
-	/* old official dataset */
-	if (ds.dbfile) {
-		/* this dataset has a db */
-		try {
-			console.log('Connecting', ds.dbfile)
-			ds.newconn = utils.connect_db(ds.dbfile)
-		} catch (e) {
-			throw 'Error with ' + ds.dbfile + ': ' + e
-		}
-	}
-
-	if (ds.snvindel_attributes) {
-		for (const at of ds.snvindel_attributes) {
-			if (at.lst) {
-				for (const a2 of at.lst) {
-					a2.get = a2.get.toString()
-				}
-			} else {
-				at.get = at.get.toString()
-			}
-		}
-	}
-
-	if (ds.cohort) {
-		// a dataset with cohort
-
-		if (ds.cohort.levels) {
-			if (!Array.isArray(ds.cohort.levels)) throw 'cohort.levels must be array for ' + ds.genomename + '.' + ds.label
-			if (ds.cohort.levels.length == 0) throw 'levels is blank array for cohort of ' + ds.genomename + '.' + ds.label
-			for (const i of ds.cohort.levels) {
-				if (!i.k) throw '.k key missing in one of the levels, .cohort, in ' + ds.genomename + '.' + ds.label
-			}
-		}
-
-		if (ds.cohort.fromdb) {
-			/*
-		cohort content to be loaded lazily from db
-		*/
-			if (!ds.cohort.fromdb.sql) throw '.sql missing from ds.cohort.fromdb in ' + ds.genomename + '.' + ds.label
-			const rows = ds.newconn.prepare(ds.cohort.fromdb.sql).all()
-			delete ds.cohort.fromdb
-			ds.cohort.raw = rows ///// backward compatible
-			console.log(rows.length + ' rows retrieved for ' + ds.label + ' sample annotation')
-		}
-
-		if (ds.cohort.files) {
-			// sample annotation load directly from text files, in sync
-			let rows = []
-			for (const file of ds.cohort.files) {
-				if (!file.file) throw '.file missing from one of cohort.files[] for ' + ds.genomename + '.' + ds.label
-				const txt = fs.readFileSync(path.join(serverconfig.tpmasterdir, file.file), 'utf8').trim()
-				if (!txt) throw file.file + ' is empty for ' + ds.genomename + '.' + ds.label
-				rows = [...rows, ...d3dsv.tsvParse(txt)]
-			}
-			delete ds.cohort.files
-			if (ds.cohort.raw) {
-				ds.cohort.raw = [...ds.cohort.raw, ...rows]
-			} else {
-				ds.cohort.raw = rows
-			}
-			console.log(rows.length + ' rows retrieved for ' + ds.label + ' sample annotation')
-		}
-		if (ds.cohort.tosampleannotation) {
-			// a directive to tell client to convert cohort.raw[] to cohort.annotation{}, key-value hash
-			if (!ds.cohort.tosampleannotation.samplekey)
-				throw '.samplekey missing from .cohort.tosampleannotation for ' + ds.genomename + '.' + ds.label
-			if (!ds.cohort.key4annotation)
-				throw (
-					'.cohort.key4annotation missing when .cohort.tosampleannotation is on for ' + ds.genomename + '.' + ds.label
-				)
-			// in fact, it still requires ds.cohort.raw, but since db querying is async, not checked
-		}
-	}
-
-	if (!ds.queries) throw '.queries missing from dataset ' + ds.label + ', ' + ds.genomename
-	if (!Array.isArray(ds.queries)) throw ds.label + '.queries is not array'
-	for (const q of ds.queries) {
-		const err = legacyds_init_one_query(q, ds, genome)
-		if (err) throw 'Error parsing a query in "' + ds.label + '": ' + err
-	}
-
-	if (ds.vcfinfofilter) {
-		const err = common.validate_vcfinfofilter(ds.vcfinfofilter)
-		if (err) throw ds.label + ': vcfinfofilter error: ' + err
-	}
-
-	if (ds.url4variant) {
-		for (const u of ds.url4variant) {
-			if (!u.makelabel) throw 'makelabel() missing for one item of url4variant from ' + ds.label
-			if (!u.makeurl) throw 'makeurl() missing for one item of url4variant from ' + ds.label
-			u.makelabel = u.makelabel.toString()
-			u.makeurl = u.makeurl.toString()
-		}
-	}
-}
-
-async function mayCreateSubdirInCache(subdir) {
-	const dir = path.join(serverconfig.cachedir, subdir)
-	try {
-		await fs.promises.stat(dir)
-	} catch (e) {
-		if (e.code == 'ENOENT') {
-			try {
-				await fs.promises.mkdir(dir)
-			} catch (e) {
-				throw 'cannot make dir'
-			}
-		} else {
-			throw 'error stating dir'
-		}
-	}
-	return dir
-}
-
-function legacyds_init_one_query(q, ds, genome) {
-	/* parse a query from legacy ds.queries[]
-	 */
-	if (!q.name) return '.name missing'
-
-	if (q.dsblocktracklst) {
-		/*
-        not sure if still in use!
-
-        one or more block track available from this query
-        quick-fix for cohort junction, replace-by-mds
-        */
-		if (!Array.isArray(q.dsblocktracklst)) return 'dsblocktracklst not an array in ' + ds.label
-		for (const tk of q.dsblocktracklst) {
-			if (!tk.type) return 'missing type for a blocktrack of ' + ds.label
-			if (!tk.file && !tk.url) return 'neither file or url given for a blocktrack of ' + ds.label
-		}
-		return
-	}
-
-	if (q.vcffile) {
-		// single vcf
-		const meta = child_process
-			.execSync(tabix + ' -H ' + path.join(serverconfig.tpmasterdir, q.vcffile), { encoding: 'utf8' })
-			.trim()
-		if (meta == '') return 'no meta lines in VCF file ' + q.vcffile + ' of query ' + q.name
-		const [info, format, samples, errs] = vcf.vcfparsemeta(meta.split('\n'))
-		if (errs) return 'error parsing VCF meta lines of ' + q.vcffile + ': ' + errs.join('; ')
-		q.vcf = {
-			vcfid: Math.random().toString(),
-			info: info,
-			format: format,
-			samples: samples
-		}
-		if (q.hlinfo) {
-			q.vcf.hlinfo = q.hlinfo
-			delete q.hlinfo
-		}
-		if (q.infopipejoin) {
-			q.vcf.infopipejoin = q.infopipejoin
-			delete q.infopipejoin
-		}
-		const tmp = child_process
-			.execSync(tabix + ' -l ' + path.join(serverconfig.tpmasterdir, q.vcffile), { encoding: 'utf8' })
-			.trim()
-		if (tmp == '') return 'tabix -l found no chromosomes/contigs in ' + q.vcffile + ' of query ' + q.name
-		q.vcf.nochr = common.contigNameNoChr(genome, tmp.split('\n'))
-		let infoc = 0
-		if (info) {
-			for (const n in info) infoc++
-		}
-		console.log(
-			'Parsed vcf meta from ' +
-				q.vcffile +
-				': ' +
-				infoc +
-				' INFO, ' +
-				samples.length +
-				' sample, ' +
-				(q.vcf.nochr ? 'no "chr"' : 'has "chr"')
-		)
-		return
-	}
-
-	if (q.makequery) {
-		if (q.isgeneexpression) {
-			if (!q.config) return 'config object missing for gene expression query of ' + q.name
-			if (q.config.maf) {
-				q.config.maf.get = q.config.maf.get.toString()
-			}
-		}
-		return
-	}
-
-	return 'do not know how to parse query: ' + q.name
-}
-
-/////////////////// __MDS
-
-async function mds_init(ds, genome, _servconfig) {
-	/*
-    ds: loaded from datasets/what.js
-    genome: obj {}
-    _servconfig: the entry in "datasets" array from serverconfig.json
-    */
-	if (ds.isMds2 || ds.isMds3 || ds.isMds) {
-		// !!! TODO: does this repeat an earlier server_updateAttr? !!!
-		server_updateAttr(ds, _servconfig)
-	} //else console.log('not mds', ds.label)
-
-	if (ds.assayAvailability) {
-		if (!ds.assayAvailability.file) throw '.assayAvailability.file missing'
-		if (!ds.assayAvailability.assays) throw '.assayAvailability.assays[] missing'
-		Object.freeze(ds.assayAvailability.assays)
-		ds.assayAvailability.samples = new Map()
-		for (const line of fs
-			.readFileSync(path.join(serverconfig.tpmasterdir, ds.assayAvailability.file), { encoding: 'utf8' })
-			.trim()
-			.split('\n')) {
-			const [sample, t] = line.split('\t')
-			ds.assayAvailability.samples.set(sample, JSON.parse(t))
-		}
-		console.log(ds.assayAvailability.samples.size + ' samples with assay availability (' + ds.label + ')')
-	}
-
-	if (ds.gene2mutcount) {
-		if (!ds.gene2mutcount.dbfile) throw '.gene2mutcount.dbfile missing'
-		try {
-			console.log('Connecting', ds.gene2mutcount.dbfile)
-			ds.gene2mutcount.db = utils.connect_db(ds.gene2mutcount.dbfile)
-			console.log('DB connected for ' + ds.label + ': ' + ds.gene2mutcount.dbfile)
-		} catch (e) {
-			throw `Error connecting db at ${ds.gene2mutcount.dbfile}` //fix for inspecific error message
-		}
-	}
-
-	if (ds.sampleAssayTrack) {
-		if (!ds.sampleAssayTrack.file) throw '.file missing from sampleAssayTrack'
-		ds.sampleAssayTrack.samples = new Map()
-
-		let count = 0
-
-		let unannotated = new Set()
-
-		for (const line of fs
-			.readFileSync(path.join(serverconfig.tpmasterdir, ds.sampleAssayTrack.file), { encoding: 'utf8' })
-			.trim()
-			.split('\n')) {
-			if (!line) continue
-			if (line[0] == '#') continue
-
-			const [sample, assay, jsontext] = line.split('\t')
-			if (!assay || !jsontext) continue
-
-			if (!ds.sampleAssayTrack.samples.has(sample)) {
-				// new sample
-				if (ds.cohort && ds.cohort.annotation) {
-					if (!ds.cohort.annotation[sample]) {
-						// this sample is unannotated
-						unannotated.add(sample)
-						continue
-					}
-				}
-
-				ds.sampleAssayTrack.samples.set(sample, [])
-			}
-
-			const tk = JSON.parse(jsontext)
-			// TODO validate track
-			if (!common.tkt[tk.type]) throw 'invalid type from a sample track: ' + jsontext
-			if (!tk.name) {
-				tk.name = sample + ' ' + assay
-			}
-
-			tk.assayName = assay
-
-			ds.sampleAssayTrack.samples.get(sample).push(tk)
-			count++
-		}
-
-		console.log(count + ' assay-tracks from ' + ds.sampleAssayTrack.samples.size + ' samples (' + ds.label + ')')
-		if (unannotated.size) {
-			console.log(
-				'Error: ' + unannotated.size + ' samples with assay tracks are unannotated: ' + [...unannotated].join(' ')
-			)
-		}
-	}
-
-	if (ds.singlesamplemutationjson) {
-		const m = ds.singlesamplemutationjson
-		if (!m.file) throw '.file missing from singlesamplemutationjson'
-		m.samples = {}
-		let count = 0
-		for (const line of fs
-			.readFileSync(path.join(serverconfig.tpmasterdir, m.file), { encoding: 'utf8' })
-			.trim()
-			.split('\n')) {
-			if (!line) continue
-			if (line[0] == '#') continue
-			const [sample, file] = line.split('\t')
-			if (sample && file) {
-				count++
-				m.samples[sample] = file
-			}
-		}
-		console.log(count + ' samples for disco plot')
-	}
-
-	if (ds.cohort && ds.cohort.db && ds.cohort.termdb) {
-		await mds2_init.init_db(ds, app, basepath)
-	}
-
-	if (ds.cohort && ds.cohort.files) {
-		/*
-        *********** legacy mds *************
-
-        following all loads sample attributes from text files
-        and store in ds.cohort.annotation
-        */
-
-		if (!Array.isArray(ds.cohort.files)) throw '.cohort.files is not array'
-
-		if (!ds.cohort.tohash) throw '.tohash() missing from cohort'
-		if (typeof ds.cohort.tohash != 'function') throw '.cohort.tohash is not function'
-		if (!ds.cohort.samplenamekey) throw '.samplenamekey missing'
-
-		// should allow both sample/individual level as key
-		ds.cohort.annotation = {}
-
-		if (ds.cohort.mutation_signature) {
-			const s = ds.cohort.mutation_signature
-			if (!s.sets) throw '.mutation_signature.sets missing'
-			for (const k in s.sets) {
-				const ss = s.sets[k]
-				if (!ss.name) ss.name = k
-				if (!ss.signatures) throw '.signatures{} missing from a signature set'
-				if (ss.samples) {
-					if (!ss.samples.file) throw '.samples.file missing from a signature set'
-					const [err, items] = parse_textfilewithheader(
-						fs.readFileSync(path.join(serverconfig.tpmasterdir, ss.samples.file), { encoding: 'utf8' }).trim()
-					)
-					ss.samples.map = new Map()
-					for (const i of items) {
-						const sample = i[ds.cohort.samplenamekey]
-						if (!sample) throw ds.cohort.samplenamekey + ' missing in file ' + ss.samples.file
-						ss.samples.map.set(sample, i)
-
-						// parse to float
-						for (const sk in ss.signatures) {
-							if (i[sk]) {
-								const v = Number.parseFloat(i[sk])
-								if (Number.isNaN(v)) throw 'mutation signature value is not float: ' + i[sk] + ' from sample ' + sample
-								if (ss.samples.skipzero && v == 0) {
-									delete i[sk]
-								} else {
-									i[sk] = v
-								}
-							}
-						}
-					}
-				}
-			}
-		}
-
-		if (ds.cohort.attributes) {
-			if (!ds.cohort.attributes.lst) throw '.lst[] missing for cohort.attributes'
-			if (!Array.isArray(ds.cohort.attributes.lst)) return '.cohort.attributes.lst is not array'
-			for (const attr of ds.cohort.attributes.lst) {
-				if (!attr.key) throw '.key missing from one of the .cohort.attributes.lst[]'
-				if (!attr.label) throw '.label missing from one of the .cohort.attributes.lst[]'
-				if (!attr.values) throw '.values{} missing from ' + attr.label + ' of .cohort.attributes.lst'
-				for (const value in attr.values) {
-					if (!attr.values[value].label)
-						throw '.label missing from one value of ' + attr.label + ' in .cohort.attributes.lst'
-				}
-			}
-			if (ds.cohort.attributes.defaulthidden) {
-				// allow attributes hidden by default
-				for (const key in ds.cohort.attributes.defaulthidden) {
-					const hideattr = ds.cohort.attributes.lst.find(i => i.key == key)
-					if (!hideattr) throw 'invalid defaulthidden key: ' + key
-					for (const value in ds.cohort.attributes.defaulthidden[key]) {
-						if (!hideattr.values[value]) throw 'invalid defaulthidden value ' + value + ' for ' + key
-					}
-				}
-			}
-		}
-
-		if (ds.cohort.hierarchies) {
-			if (!ds.cohort.hierarchies.lst) throw '.lst[] missing from .cohort.hierarchies'
-			if (!Array.isArray(ds.cohort.hierarchies.lst)) throw '.cohort.hierarchies.lst is not array'
-			for (const h of ds.cohort.hierarchies.lst) {
-				if (!h.name) throw '.name missing from one hierarchy'
-				if (!h.levels) throw '.levels[] missing from one hierarchy'
-				if (!Array.isArray(h.levels)) throw '.levels is not array from one hierarchy'
-				for (const l of h.levels) {
-					if (!l.k) throw '.k missing from one level in hierarchy ' + h.name
-				}
-			}
-		}
-
-		if (ds.cohort.sampleAttribute) {
-			if (!ds.cohort.sampleAttribute.attributes) throw 'attributes{} missing from cohort.sampleAttribute'
-			for (const key in ds.cohort.sampleAttribute.attributes) {
-				const a = ds.cohort.sampleAttribute.attributes[key]
-				if (!a.label) throw '.label missing for key ' + key + ' from cohort.sampleAttribute.attributes'
-				if (a.values) {
-					// optional
-					for (const v in a.values) {
-						const b = a.values[v]
-						if (typeof b != 'object') throw 'value "' + v + '" not pointing to {} from sampleAttribute'
-						if (!b.name) b.name = v
-					}
-				}
-				if (a.showintrack) {
-					if (!a.isinteger && !a.isfloat) throw a.label + ': .showintrack requires .isinteger or .isfloat'
-				}
-			}
-		}
-
-		if (ds.cohort.scatterplot) {
-			if (!ds.cohort.sampleAttribute) throw '.sampleAttribute missing but required for .cohort.scatterplot'
-
-			const sp = ds.cohort.scatterplot
-
-			// querykey is required
-			if (!sp.querykey) throw '.querykey missing from .cohort.scatterplot'
-			{
-				if (!ds.queries) throw '.cohort.scatterplot.querykey in use but ds.queries{} missing'
-				const tk = ds.queries[sp.querykey]
-				if (!tk) throw 'unknown query by .cohort.scatterplot.querykey: ' + sp.querykey
-				if (tk.type != common.tkt.mdssvcnv)
-					throw 'type is not ' + common.tkt.mdssvcnv + ' of the track pointed to by .cohort.scatterplot.querykey'
-			}
-
-			if (sp.colorbygeneexpression) {
-				if (!sp.colorbygeneexpression.querykey) throw 'querykey missing from .cohort.scatterplot.colorbygeneexpression'
-				if (!ds.queries) throw '.cohort.scatterplot.colorbygeneexpression in use but ds.queries{} missing'
-				const tk = ds.queries[sp.colorbygeneexpression.querykey]
-				if (!tk)
-					throw (
-						'unknown query by .cohort.scatterplot.colorbygeneexpression.querykey: ' + sp.colorbygeneexpression.querykey
-					)
-				if (!tk.isgenenumeric)
-					throw 'isgenenumeric missing from the track pointed to by .cohort.scatterplot.colorbygeneexpression.querykey'
-			}
-
-			if (sp.tracks) {
-				// a common set of tracks to be shown in single sample browser upon clicking a dot
-				// must label them as custom otherwise they won't be listed
-				// TODO validate the tracks
-			}
-
-			// TODO support multiple plots
-			if (!sp.x) throw '.x missing from .cohort.scatterplot'
-			if (!sp.x.attribute) throw '.attribute missing from .cohort.scatterplot.x'
-			const x = ds.cohort.sampleAttribute.attributes[sp.x.attribute]
-			if (!x) throw 'scatterplot.x.attribute is not defined in sampleAttribute'
-			if (!x.isfloat) throw 'scatterplot.x is not "isfloat"'
-			if (!sp.y) throw '.y missing from .cohort.scatterplot'
-			if (!sp.y.attribute) throw '.attribute missing from .cohort.scatterplot.y'
-			const y = ds.cohort.sampleAttribute.attributes[sp.y.attribute]
-			if (!y) throw 'scatterplot.y.attribute is not defined in sampleAttribute'
-			if (!y.isfloat) throw 'scatterplot.y is not "isfloat"'
-			if (sp.colorbyattributes) {
-				for (const attr of sp.colorbyattributes) {
-					if (!attr.key) throw '.key missing from one of scatterplot.colorbyattributes'
-					const attrreg = ds.cohort.sampleAttribute.attributes[attr.key]
-					if (!attrreg) throw 'unknown attribute by key ' + attr.key + ' from scatterplot.colorbyattributes'
-					attr.label = attrreg.label
-					attr.values = attrreg.values
-				}
-			}
-		}
-
-		for (const file of ds.cohort.files) {
-			if (!file.file) throw '.file missing from one of .cohort.files'
-			const [err, items] = parse_textfilewithheader(
-				fs.readFileSync(path.join(serverconfig.tpmasterdir, file.file), { encoding: 'utf8' }).trim()
-			)
-			if (err) throw 'cohort annotation file "' + file.file + '": ' + err
-			//if(items.length==0) return 'no content from sample annotation file '+file.file
-			console.log(ds.label + ': ' + items.length + ' samples loaded from annotation file ' + file.file)
-			items.forEach(i => {
-				// may need to parse certain values into particular format
-
-				for (const k in i) {
-					let attr
-					if (ds.cohort.sampleAttribute) {
-						attr = ds.cohort.sampleAttribute.attributes[k]
-					}
-					if (!attr) {
-						if (ds.cohort.termdb && ds.cohort.termdb.termjson && ds.cohort.termdb.termjson.map) {
-							attr = ds.cohort.termdb.termjson.map.get(k)
-						}
-					}
-					if (attr) {
-						if (attr.isfloat) {
-							i[k] = Number.parseFloat(i[k])
-						} else if (attr.isinteger) {
-							i[k] = Number.parseInt(i[k])
-						}
-					}
-				}
-
-				ds.cohort.tohash(i, ds)
-			})
-		}
-		ds.cohort.annorows = Object.values(ds.cohort.annotation)
-		console.log(ds.label + ': total samples from sample table: ' + ds.cohort.annorows.length)
-
-		if (ds.cohort.survivalplot) {
-			// ds.cohort.annotation needs to be loaded for initing survival
-			const sp = ds.cohort.survivalplot
-			if (!sp.plots) throw '.plots{} missing from survivalplot'
-
-			// make the object for initiating client
-			sp.init = {
-				plottypes: []
-			}
-
-			for (const k in sp.plots) {
-				const p = sp.plots[k]
-				if (!p.name) throw '.name missing from survivalplot ' + k
-				if (!p.serialtimekey) throw '.serialtimekey missing from survivalplot ' + k
-				if (!p.iscensoredkey) throw '.iscensoredkey missing from survivalplot ' + k
-				if (!p.timelabel) throw '.timelabel missing from survivalplot ' + k
-				p.key = k
-
-				sp.init.plottypes.push({
-					key: k,
-					name: p.name,
-					timelabel: p.timelabel
-				})
-			}
-
-			if (sp.samplegroupattrlst) {
-				sp.init.samplegroupings = []
-				for (const a of sp.samplegroupattrlst) {
-					if (!a.key) throw '.key missing from an attr of samplegroupattrlst for survival'
-
-					const attr = ds.cohort.sampleAttribute.attributes[a.key]
-					if (!attr) throw 'unknown attribute key "' + a.key + '" from survival samplegroupattrlst'
-
-					const value2count = new Map()
-					for (const samplename in ds.cohort.annotation) {
-						const sobj = ds.cohort.annotation[samplename]
-						const v = sobj[a.key]
-						if (v == undefined) {
-							// sample not annotated by it
-							continue
-						}
-
-						let hasoutcome = false
-						// if the sample has info in any plot, will count it
-						for (const k in sp.plots) {
-							if (sobj[sp.plots[k].serialtimekey] != undefined) {
-								hasoutcome = true
-								break
-							}
-						}
-
-						if (hasoutcome) {
-							if (value2count.has(v)) {
-								value2count.set(v, value2count.get(v) + 1)
-							} else {
-								value2count.set(v, 1)
-							}
-						}
-					}
-					if (value2count.size == 0) throw 'no value found for "' + a.key + '" from survival samplegroupattrlst'
-
-					const lst = []
-					for (const [v, c] of value2count) {
-						lst.push({ value: v, count: c })
-					}
-					sp.init.samplegroupings.push({
-						key: a.key,
-						label: attr.label,
-						values: lst
-					})
-				}
-			}
-		}
-	}
-
-	if (ds.mutationAttribute) {
-		/*
-        mutation-level attributes
-        for items in svcnv track:
-        	.mattr{}
-        for vcf:
-        	FORMAT
-        */
-		if (!ds.mutationAttribute.attributes) throw 'attributes{} missing from mutationAttribute'
-		for (const key in ds.mutationAttribute.attributes) {
-			const a = ds.mutationAttribute.attributes[key]
-			if (!a.label) throw '.label missing for key ' + key + ' from mutationAttribute.attributes'
-			if (a.appendto_link) {
-				// this is pmid, no .values{}
-				continue
-			}
-			if (a.values) {
-				for (const v in a.values) {
-					const b = a.values[v]
-					if (!b.name) throw '.name missing for value ' + v + ' of key ' + key + ' from mutationAttribute.attributes'
-				}
-			} else {
-				// allow values{} missing
-			}
-		}
-	}
-
-	if (ds.alleleAttribute) {
-		/*
-        vcf info field, allele-level
-        */
-		if (!ds.alleleAttribute.attributes) throw 'attributes{} missing from alleleAttribute'
-		for (const key in ds.alleleAttribute.attributes) {
-			const a = ds.alleleAttribute.attributes[key]
-			if (!a.label) throw '.label missing for key ' + key + ' from alleleAttribute.attributes'
-			if (a.isnumeric) {
-				continue
-			}
-			// not numeric value
-			if (!a.values) throw '.values{} missing for non-numeric key ' + key + ' from alleleAttribute.attributes'
-			for (const v in a.values) {
-				const b = a.values[v]
-				if (!b.name) throw '.name missing for value ' + v + ' of key ' + key + ' from alleleAttribute.attributes'
-			}
-		}
-	}
-
-	if (ds.locusAttribute) {
-		/*
-        vcf info field, locus-level
-        */
-		if (!ds.locusAttribute.attributes) throw 'attributes{} missing from locusAttribute'
-		for (const key in ds.locusAttribute.attributes) {
-			const a = ds.locusAttribute.attributes[key]
-			if (!a.label) throw '.label missing for key ' + key + ' from locusAttribute.attributes'
-			if (a.isnumeric) {
-				continue
-			}
-			if (a.appendto_link) {
-				// no .values{}
-				continue
-			}
-			// not numeric value
-			if (!a.values) throw '.values{} missing for non-numeric key ' + key + ' from locusAttribute.attributes'
-			for (const v in a.values) {
-				const b = a.values[v]
-				if (!b.name) throw '.name missing for value ' + v + ' of key ' + key + ' from locusAttribute.attributes'
-			}
-		}
-	}
-
-	if (ds.queries) {
-		// ds.queries is the 1st generation track
-
-		for (const querykey in ds.queries) {
-			// server may choose to remove it
-			if (_servconfig.remove_queries && _servconfig.remove_queries.indexOf(querykey) != -1) {
-				delete ds.queries[querykey]
-				continue
-			}
-
-			const query = ds.queries[querykey]
-
-			// server may choose to hide some queries
-			if (_servconfig.hide_queries && _servconfig.hide_queries.indexOf(querykey) != -1) {
-				// this query will be hidden on client
-				query.hideforthemoment = 1
-			}
-
-			if (query.istrack) {
-				if (!query.type) throw '.type missing for track query ' + querykey
-
-				if (query.viewrangeupperlimit) {
-					if (!Number.isInteger(query.viewrangeupperlimit))
-						throw '.viewrangeupperlimit should be integer for track query ' + querykey
-				}
-
-				if (query.type == common.tkt.mdsjunction) {
-					const err = mds_init_mdsjunction(query, ds, genome)
-					if (err) throw querykey + ' (mdsjunction) error: ' + err
-				} else if (query.type == common.tkt.mdscnv) {
-					// obsolete
-
-					const err = mds_init_mdscnv(query, ds, genome)
-					if (err) throw querykey + ' (mdscnv) error: ' + err
-				} else if (query.type == common.tkt.mdssvcnv) {
-					// replaces mdscnv
-
-					const err = mds_init_mdssvcnv(query, ds, genome)
-					if (err) throw querykey + ' (svcnv) error: ' + err
-				} else if (query.type == common.tkt.mdsvcf) {
-					// snvindel
-
-					const err = await mds_init_mdsvcf(query, ds, genome)
-					if (err) throw querykey + ' (vcf) error: ' + err
-				} else {
-					throw 'unknown track type for a query: ' + query.type + ' ' + querykey
-				}
-
-				mds_mayPrecompute_grouptotal(query, ds)
-			} else if (query.isgenenumeric) {
-				const err = mds_init_genenumeric(query, ds, genome)
-				if (err) throw querykey + ' (genenumeric) error: ' + err
-			} else {
-				throw 'unknown type of query from ' + querykey
-			}
-		}
-	}
-
-	if (ds.track) {
-		await mds2_init.init_track(ds, genome)
-	}
-
-	if (ds.annotationsampleset2matrix) {
-		if (!ds.cohort) throw 'ds.cohort misssing when annotationsampleset2matrix is in use'
-		if (!ds.cohort.annotation) throw 'ds.cohort.annotation misssing when annotationsampleset2matrix is in use'
-		if (!ds.queries) throw 'ds.queries misssing when annotationsampleset2matrix is in use'
-		if (!ds.annotationsampleset2matrix.key) throw '.key STR missing in annotationsampleset2matrix'
-		if (!ds.annotationsampleset2matrix.groups) throw '.groups{} missing in annotationsampleset2matrix'
-		if (typeof ds.annotationsampleset2matrix.groups != 'object')
-			throw 'ds.annotationsampleset2matrix.groups{} not an object'
-		for (const groupvalue in ds.annotationsampleset2matrix.groups) {
-			const thisgroup = ds.annotationsampleset2matrix.groups[groupvalue]
-			// a group will have 1 or more smaller groups, each is samples from a study
-			if (!thisgroup.groups || !Array.isArray(thisgroup.groups) || thisgroup.groups.length == 0)
-				throw '.groups[] must be nonempty array in ' + groupvalue
-			for (const group of thisgroup.groups) {
-				if (!group.name) throw '.name missing from one of .groups[] in ' + groupvalue
-				const smat = group.matrixconfig
-				if (!smat) throw '.matrixconfig missing from one of .groups[] of ' + groupvalue
-				if (!smat.features) throw '.features[] missing from group ' + groupvalue
-				if (!Array.isArray(smat.features)) throw '.features[] should be array from group ' + groupvalue
-				if (smat.features.length == 0) throw '.features[] zero length from group ' + groupvalue
-				for (const feature of smat.features) {
-					if (ds.annotationsampleset2matrix.commonfeatureattributes) {
-						// apply common attributes to each feature
-						for (const k in ds.annotationsampleset2matrix.commonfeatureattributes) {
-							// not to overwrite predefined value
-							if (feature[k] == undefined) {
-								feature[k] = ds.annotationsampleset2matrix.commonfeatureattributes[k]
-							}
-						}
-					}
-					if (feature.ismutation) {
-						if (!feature.position)
-							throw 'position missing from feature ' + JSON.stringify(feature) + ' from group ' + groupvalue
-						if (!feature.querykeylst) throw '.querykeylst[] missing from ismutation feature from group ' + groupvalue
-						if (!Array.isArray(feature.querykeylst))
-							throw '.querykeylst[] not an array from ismutation feature from group ' + groupvalue
-						if (feature.querykeylst.length == 0)
-							throw '.querykeylst[] zero length from ismutation feature from group ' + groupvalue
-						for (const querykey of feature.querykeylst) {
-							if (!ds.queries[querykey])
-								throw 'unknown query key "' + querykey + '" from ismutation feature of group ' + groupvalue
-						}
-						continue
-					}
-					return 'unknown feature type from group ' + groupvalue
-				}
-				if (!smat.limitsamplebyeitherannotation)
-					throw '.limitsamplebyeitherannotation[] missing from group ' + groupvalue
-				if (!Array.isArray(smat.limitsamplebyeitherannotation))
-					throw '.limitsamplebyeitherannotation[] should be array from group ' + groupvalue
-				if (smat.limitsamplebyeitherannotation.length == 0)
-					throw '.limitsamplebyeitherannotation[] zero length from group ' + groupvalue
-				for (const lim of smat.limitsamplebyeitherannotation) {
-					if (!lim.key) throw 'key missing from one of limitsamplebyeitherannotation from group ' + groupvalue
-					if (!lim.value) throw 'value missing from one of limitsamplebyeitherannotation from group ' + groupvalue
-				}
-			}
-		}
-		delete ds.annotationsampleset2matrix.commonfeatureattributes
-	}
-}
-
-function mds_mayPrecompute_grouptotal(query, ds) {
-	if (!query.groupsamplebyattr) return
-	query.groupsamplebyattr.key2group = new Map()
-	for (const samplename in ds.cohort.annotation) {
-		mdssvcnv_grouper(samplename, [], query.groupsamplebyattr.key2group, [], ds, query)
-	}
-}
-
-function mds_init_mdsjunction(query, ds, genome) {
-	// mdsjunction only allows single track
-
-	if (query.readcountCutoff) {
-		if (!Number.isInteger(query.readcountCutoff) || query.readcountCutoff < 1)
-			return 'readcountCutoff must be positive integer'
-	}
-
-	let cwd = null
-	let _file
-
-	if (query.file) {
-		const [err, tmp] = validate_tabixfile(query.file)
-		if (err) return 'tabix file error: ' + err
-		query.file = tmp // replace with full path
-		_file = tmp
-	} else if (query.url) {
-		// TODO cache_index sync
-		// need to set query.usedir to cache path
-		_file = query.url
-	} else {
-		return 'no file or url given for mdsjunction ' + query.name
-	}
-
-	const arg = { encoding: 'utf8' }
-	if (cwd) {
-		arg.cwd = cwd
-	}
-
-	const header = child_process.execSync(tabix + ' -H ' + _file, arg).trim()
-	if (header) {
-		// has header, get samples
-		const lines = header.split('\n')
-		if (lines.length != 1) return 'mdsjunction file has multiple header lines (begin with #), but should have just 1'
-		const lst = lines[0].split('\t')
-		// #chr \t start \t stop \t strand \t type \t samples ...
-		if (lst[5]) {
-			query.samples = lst.slice(5)
-			query.attributeSummary = mds_query_attrsum4samples(query.samples, ds)
-			query.hierarchySummary = mds_query_hierarchy4samples(query.samples, ds)
-			for (const name in query.hierarchySummary) {
-				let levelcount = 0
-				for (const k in query.hierarchySummary[name]) levelcount++
-				console.log(levelcount + ' ' + name + ' hierarchy levels for ' + query.name)
-			}
-		}
-	}
-
-	{
-		const tmp = child_process.execSync(tabix + ' -l ' + _file, arg).trim()
-		if (!tmp) return 'no chromosomes found'
-		query.nochr = common.contigNameNoChr(genome, tmp.split('\n'))
-	}
-
-	console.log(
-		'(mdsjunction) ' +
-			query.name +
-			': ' +
-			(query.samples ? query.samples.length : 0) +
-			' samples, ' +
-			(query.nochr ? 'no "chr"' : 'has "chr"')
-	)
-
-	if (!query.infoFilter) return '.infoFilter{} missing'
-	if (!query.infoFilter.lst) return '.lst[] missing from .infoFilter'
-	// currently infoFilter contains Type (column 5) and splice events, both are categorical
-	for (const info of query.infoFilter.lst) {
-		if (!info.key) return '.key missing from one of infoFilter'
-		if (!info.label) return '.label missing from one of infoFilter'
-		if (!info.categories) return '.categories missing from one of infoFilter'
-		for (const k in info.categories) {
-			if (!info.categories[k].label) return '.label missing from one category of ' + info.label
-			if (!info.categories[k].color) return '.color missing from on category of ' + info.label
-		}
-		if (info.hiddenCategories) {
-			// allow initially hidden categories
-			for (const k in info.hiddenCategories) {
-				if (!info.categories[k]) return 'invalid hidden key ' + k + ' of ' + info.label
-			}
-		} else {
-			info.hiddenCategories = {}
-		}
-	}
-
-	if (!query.singlejunctionsummary) return '.singlejunctionsummary missing but is currently required from ' + query.name
-	if (query.singlejunctionsummary.readcountboxplotpercohort) {
-		if (!query.singlejunctionsummary.readcountboxplotpercohort.groups)
-			return '.groups[] missing from query.singlejunctionsummary.readcountboxplotpercohort for ' + query.name
-		for (const g of query.singlejunctionsummary.readcountboxplotpercohort.groups) {
-			if (!g.key) return '.key missing from one group of query.singlejunctionsummary.readcountboxplotpercohort.groups'
-			if (!g.label)
-				return '.label missing from one group of query.singlejunctionsummary.readcountboxplotpercohort.groups'
-		}
-	}
-}
-
-function mds_query_attrsum4samples(samples, ds) {
-	/*
-    summarizes a group of samples by list of attributes in ds.cohort.attributes.lst[]
-
-    a query from mds has total list of samples, e.g. samples in mdsjunction represent those with RNA-seq
-    for these samples, will sum up .totalCount for cohort annotation attributes/values (by ds.cohort.attributes)
-    rather than computing .totalCount over all samples of the ds.cohort, so as to limit to relevant assays
-    so on cohortFilter legend it will only show totalCount from those samples with RNA-seq etc
-    */
-	if (!ds.cohort || !ds.cohort.annotation || !ds.cohort.attributes || !samples) return
-
-	const result = {}
-	for (const attr of ds.cohort.attributes.lst) {
-		// TODO numeric attribute?
-
-		const v2c = {}
-		for (const value in attr.values) {
-			v2c[value] = 0
-		}
-		// go over samples look for this attribute
-		for (const sample of samples) {
-			const anno = ds.cohort.annotation[sample]
-			if (!anno) {
-				// this sample has no annotation
-				continue
-			}
-			const thisvalue = anno[attr.key]
-			if (thisvalue == undefined) {
-				// this sample not annotated by this attribute
-				continue
-			}
-			if (thisvalue in v2c) {
-				v2c[thisvalue]++
-			}
-		}
-		result[attr.key] = v2c
-	}
-	return result
-}
-
-function mds_query_hierarchy4samples(samples, ds) {
-	/*
-    given a list of sample names, generate hierarchy summary
-
-    	key: hierarchy path (HM...BALL...ERG)
-    	value: number of samples
-
-    works for both initializing the sample sets from each ds query, and also for samples in view range in real-time track query
-    */
-	if (!ds.cohort || !ds.cohort.annotation || !ds.cohort.hierarchies || samples.length == 0) return
-	const lst = []
-	for (const n of samples) {
-		const a = ds.cohort.annotation[n]
-		if (!a) continue
-		lst.push(a)
-	}
-
-	const results = {}
-
-	for (const hierarchy of ds.cohort.hierarchies.lst) {
-		const nodes = stratinput(lst, hierarchy.levels)
-		const root = d3stratify()(nodes)
-		root.sum(i => i.value)
-		const id2count = {}
-		root.eachBefore(i => {
-			id2count[i.id] = i.value
-		})
-		results[hierarchy.name] = id2count
-	}
-	return results
-}
-
-function mds_init_mdscnv(query, ds, genome) {
-	// mdscnv only allows single track
-
-	let cwd = null
-	let _file
-
-	if (query.file) {
-		const [err, tmp] = validate_tabixfile(query.file)
-		if (err) return 'tabix file error: ' + err
-		_file = tmp
-	} else if (query.url) {
-		// TODO cache_index sync
-		// need to set query.usedir to cache path
-		_file = query.url
-	} else {
-		return 'no file or url given for (mdscnv) ' + query.name
-	}
-
-	const arg = { encoding: 'utf8' }
-	if (cwd) {
-		arg.cwd = cwd
-	}
-
-	const header = child_process.execSync(tabix + ' -H ' + _file, arg).trim()
-	if (header) {
-		// has header, get samples
-		const lines = header.split('\n')
-		if (lines.length != 1) return 'mdscnv file has multiple header lines (begin with #), but should have just 1'
-		const lst = lines[0].split('\t')
-		query.samples = lst.slice(5)
-		query.attributeSummary = mds_query_attrsum4samples(query.samples, ds)
-		query.hierarchySummary = mds_query_hierarchy4samples(query.samples, ds)
-		for (const name in query.hierarchySummary) {
-			let levelcount = 0
-			for (const k in query.hierarchySummary[name]) levelcount++
-			console.log(levelcount + ' ' + name + ' hierarchy levels for ' + query.name)
-		}
-	}
-
-	{
-		const tmp = child_process.execSync(tabix + ' -l ' + _file, arg).trim()
-		if (!tmp) return 'no chromosomes found'
-		query.nochr = common.contigNameNoChr(genome, tmp.split('\n'))
-	}
-
-	console.log(
-		'(' +
-			query.type +
-			') ' +
-			query.name +
-			': ' +
-			(query.samples ? query.samples.length : 'no') +
-			' samples, ' +
-			(query.nochr ? 'no "chr"' : 'has "chr"')
-	)
-}
-
-function mds_init_mdssvcnv(query, ds, genome) {
-	// only allows single track, since there is no challenge merging multiple into one
-
-	let cwd = null
-	let _file
-
-	if (query.file) {
-		const [err, tmp] = validate_tabixfile(query.file)
-		if (err) return 'tabix file error: ' + err
-		_file = tmp
-	} else if (query.url) {
-		// TODO cache_index sync
-		// need to set query.usedir to cache path
-		_file = query.url
-	} else {
-		return 'no file or url given for (svcnv) ' + query.name
-	}
-
-	const arg = { encoding: 'utf8' }
-	if (cwd) {
-		arg.cwd = cwd
-	}
-
-	const header = child_process.execSync(tabix + ' -H ' + _file, arg).trim()
-	if (header) {
-		// has header, get samples
-		const set = new Set()
-		for (const line of header.split('\n')) {
-			for (const s of line.split(' ').slice(1)) {
-				set.add(s)
-			}
-		}
-		if (set.size == 0) return 'no samples from the header line'
-		query.samples = [...set]
-
-		if (ds.cohort && ds.cohort.annotation) {
-			// find & report unannotated samples
-			const unknown = new Set()
-			for (const sample of query.samples) {
-				if (!ds.cohort.annotation[sample]) {
-					unknown.add(sample)
-				}
-			}
-			if (unknown.size) {
-				console.log(
-					'mdssvcnv unannotated samples: ' + (query.noprintunannotatedsamples ? unknown.size : [...unknown].join(' '))
-				)
-			}
-		}
-
-		/*
-        // not used at the moment
-        query.attributeSummary = mds_query_attrsum4samples(query.samples, ds)
-        query.hierarchySummary = mds_query_hierarchy4samples(query.samples,ds)
-        for(const hierarchyname in query.hierarchySummary) {
-        	let levelcount=0
-        	for(const k in query.hierarchySummary[ hierarchyname ]) levelcount++
-        	console.log(levelcount+' '+hierarchyname+' hierarchy levels for '+query.name)
-        }
-        */
-	}
-
-	{
-		const tmp = child_process.execSync(tabix + ' -l ' + _file, arg).trim()
-		if (!tmp) return 'no chromosomes found'
-		query.nochr = common.contigNameNoChr(genome, tmp.split('\n'))
-	}
-
-	if (query.expressionrank_querykey) {
-		// check expression rank, data from another query
-		const thatquery = ds.queries[query.expressionrank_querykey]
-		if (!thatquery) return 'invalid key by expressionrank_querykey'
-		if (!thatquery.isgenenumeric) return 'query ' + query.expressionrank_querykey + ' not tagged as isgenenumeric'
-	}
-
-	if (query.vcf_querykey) {
-		// check expression rank, data from another query
-		const thatquery = ds.queries[query.vcf_querykey]
-		if (!thatquery) return 'invalid key by vcf_querykey'
-		if (thatquery.type != common.tkt.mdsvcf) return 'query ' + query.vcf_querykey + ' not of mdsvcf type'
-	}
-
-	if (query.groupsamplebyattr) {
-		if (!query.groupsamplebyattr.attrlst) return '.attrlst[] missing from groupsamplebyattr'
-		if (query.groupsamplebyattr.attrlst.length == 0) return 'groupsamplebyattr.attrlst[] empty array'
-		if (!ds.cohort) return 'groupsamplebyattr in use but ds.cohort missing'
-		if (!ds.cohort.annotation) return 'groupsamplebyattr in use but ds.cohort.annotation missing'
-		if (!ds.cohort.sampleAttribute) {
-			ds.cohort.sampleAttribute = {}
-		}
-		if (!ds.cohort.sampleAttribute.attributes) {
-			ds.cohort.sampleAttribute.attributes = {}
-			console.log('cohort.sampleAttribute added when groupsamplebyattr is in use')
-		}
-		for (const attr of query.groupsamplebyattr.attrlst) {
-			if (!attr.k) return 'k missing from one of groupsamplebyattr.attrlst[]'
-			if (!ds.cohort.sampleAttribute.attributes[attr.k]) {
-				ds.cohort.sampleAttribute.attributes[attr.k] = {
-					label: attr.label || attr.k
-				}
-			}
-		}
-		if (query.groupsamplebyattr.sortgroupby) {
-			if (!query.groupsamplebyattr.sortgroupby.key) return '.key missing from .sortgroupby'
-			if (!query.groupsamplebyattr.sortgroupby.order) return '.order[] missing from .sortgroupby'
-			if (!Array.isArray(query.groupsamplebyattr.sortgroupby.order)) return '.order must be an array'
-			// values of order[] is not validated
-		}
-		if (!query.groupsamplebyattr.attrnamespacer) query.groupsamplebyattr.attrnamespacer = ', '
-	}
-
-	console.log(
-		'(' +
-			query.type +
-			') ' +
-			query.name +
-			': ' +
-			(query.samples ? query.samples.length : 'no') +
-			' samples, ' +
-			(query.nochr ? 'no "chr"' : 'has "chr"')
-	)
-}
-
-function mds_init_genenumeric(query, ds, genome) {
-	if (!query.datatype) return 'datatype missing'
-	if (query.viewrangeupperlimit) {
-		if (Number.isNaN(query.viewrangeupperlimit)) return 'invalid value for viewrangeupperlimit'
-	}
-
-	let cwd = null
-	let _file
-	if (query.file) {
-		const [err, tmp] = validate_tabixfile(query.file)
-		if (err) return 'tabix file error: ' + err
-		_file = tmp
-	} else {
-		// no url support yet
-		return 'file missing'
-	}
-
-	const arg = { cwd: cwd, encoding: 'utf8' }
-
-	{
-		const tmp = child_process.execSync(tabix + ' -H ' + _file, arg).trim()
-		if (!tmp) return 'no header line (#sample <sample1> ...)'
-		// allow multiple # lines
-		const set = new Set()
-		for (const line of tmp.split('\n')) {
-			const l = line.split(' ')
-			for (let i = 1; i < l.length; i++) {
-				set.add(l[i])
-			}
-		}
-		if (set.size == 0) return 'no sample names from header line'
-		query.samples = [...set]
-		console.log('(genenumeric) ' + query.name + ': ' + query.samples.length + ' samples')
-	}
-	if (query.boxplotbysamplegroup) {
-		if (!query.boxplotbysamplegroup.attributes) return 'boxplotbysamplegroup.attributes missing'
-		if (!Array.isArray(query.boxplotbysamplegroup.attributes)) return 'boxplotbysamplegroup.attributes should be array'
-		for (const a of query.boxplotbysamplegroup.attributes) {
-			if (!a.k) return 'k missing from one of boxplotbysamplegroup.attributes[]'
-		}
-	}
-}
-
-async function mds_init_mdsvcf(query, ds, genome) {
-	/*
-    mixture of snv/indel (vcf), ITD, and others
-    that are not either cnv or sv
-    has member tracks, each track of one type of data
-    */
-
-	if (!query.tracks) return 'tracks[] missing'
-	if (!Array.isArray(query.tracks)) return 'tracks should be array'
-
-	/*
-    info from all member tracks are merged, this requires the same info shared across multiple tracks must be identical
-    */
-	query.info = {}
-
-	for (const tk of query.tracks) {
-		if (!tk.file) return 'file missing from a track (url not supported yet)'
-
-		// will set tk.cwd for url
-
-		const [err, _file] = validate_tabixfile(tk.file)
-		if (err) return 'tabix file error: ' + err
-
-		if (tk.type == common.mdsvcftype.vcf) {
-			const arg = { cwd: tk.cwd, encoding: 'utf8' }
-
-			const tmp = await utils.get_header_tabix(_file, tk.cwd)
-			if (tmp.length == 0) return 'no meta/header lines for ' + _file
-			const [info, format, samples, errs] = vcf.vcfparsemeta(tmp)
-			if (errs) return 'error parsing vcf meta for ' + _file + ': ' + errs.join('\n')
-
-			if (samples.length == 0) return 'vcf file has no sample: ' + _file
-
-			for (const k in info) {
-				query.info[k] = info[k]
-			}
-
-			tk.format = format
-
-			if (tk.samplenameconvert) {
-				if (typeof tk.samplenameconvert != 'function') return '.samplenameconvert must be function'
-				for (const s of samples) {
-					s.name = tk.samplenameconvert(s.name)
-				}
-			}
-
-			tk.samples = samples
-		} else {
-			return 'invalid track type: ' + tk.type
-		}
-
-		if (ds.cohort && ds.cohort.annotation) {
-			/*
-            ds.cohort.annotation is sample-level, e.g. tumor
-            if vcf encodes germline stuff on person, or need some kind of sample name conversion,
-            need to identify such in this track
-            */
-			const notannotated = []
-			for (const s of tk.samples) {
-				if (!ds.cohort.annotation[s.name]) {
-					notannotated.push(s.name)
-				}
-			}
-			if (notannotated.length) {
-				console.log(ds.label + ': VCF ' + tk.file + ' has unannotated samples: ' + notannotated.join(','))
-			}
-		}
-
-		{
-			const tmp = []
-			await utils.get_lines_bigfile({
-				args: ['-l', _file],
-				dir: tk.cwd,
-				callback: line => {
-					tmp.push(line)
-				}
-			})
-			if (tmp.length == 0) return 'no chr from ' + _file
-			tk.nochr = common.contigNameNoChr(genome, tmp)
-		}
-
-		console.log(
-			'(' + query.type + ') ' + _file + ': ' + tk.samples.length + ' samples, ' + (tk.nochr ? 'no chr' : 'has chr')
-		)
-	}
-
-	if (query.singlesamples) {
-		if (!query.singlesamples.tablefile) return '.singlesamples.tablefile missing for the VCF query'
-		query.singlesamples.samples = {}
-		let count = 0
-		for (const line of fs
-			.readFileSync(path.join(serverconfig.tpmasterdir, query.singlesamples.tablefile), { encoding: 'utf8' })
-			.trim()
-			.split('\n')) {
-			if (!line) continue
-			if (line[0] == '#') continue
-			const l = line.split('\t')
-			if (l[0] && l[1]) {
-				query.singlesamples.samples[l[0]] = l[1]
-				count++
-			}
-		}
-		console.log(count + ' single-sample VCF files')
-	}
-}
-
-////////////// end of __MDS
 
 function handle_ideogram(req, res) {
 	try {
