+++ server/src/bam.js
@@ -1,20 +1,20 @@
 import fs from 'fs'
 import path from 'path'
-import * as utils from './utils'
-import serverconfig from './serverconfig'
+import * as utils from './utils.js'
+import serverconfig from './serverconfig.js'
 import { spawn } from 'child_process'
 import { Readable, Transform } from 'stream'
 import { pipeline } from 'node:stream/promises'
 import { createCanvas } from 'canvas'
-import * as bamcommon from './bam.common'
+import * as bamcommon from './bam.common.js'
 import { run_rust } from '@sjcrh/proteinpaint-rust'
 import crypto from 'crypto'
 import got from 'got'
 import { interpolateRgb } from 'd3-interpolate'
-import { match_complexvariant_rust } from './bam.kmer.indel'
-import { basecolor, bplen } from '#shared/common'
-import { gdcCheckPermission } from './bam.gdc'
-import { fileSize } from '../shared/fileSize'
+import { match_complexvariant_rust } from './bam.kmer.indel.js'
+import { basecolor, bplen } from '#shared/common.js'
+import { gdcCheckPermission } from './bam.gdc.js'
+import { fileSize } from '../shared/fileSize.js'
 
 const clustalo_read_alignment = serverconfig.clustalo
