+++ server/src/handle_study.js
@@ -1,16 +1,16 @@
 import path from 'path'
-import serverconfig from './serverconfig'
-import * as utils from './utils'
-import * as common from '#shared/common'
-import * as vcf from '#shared/vcf'
-import * as bulk from '#shared/bulk'
-import * as bulksnv from '#shared/bulk.snv'
-import * as bulkcnv from '#shared/bulk.cnv'
-import * as bulkdel from '#shared/bulk.del'
-import * as bulkitd from '#shared/bulk.itd'
-import * as bulksv from '#shared/bulk.sv'
-import * as bulksvjson from '#shared/bulk.svjson'
-import * as bulktrunc from '#shared/bulk.trunc'
+import serverconfig from './serverconfig.js'
+import * as utils from './utils.js'
+import * as common from '#shared/common.js'
+import * as vcf from '#shared/vcf.js'
+import * as bulk from '#shared/bulk.js'
+import * as bulksnv from '#shared/bulk.snv.js'
+import * as bulkcnv from '#shared/bulk.cnv.js'
+import * as bulkdel from '#shared/bulk.del.js'
+import * as bulkitd from '#shared/bulk.itd.js'
+import * as bulksv from '#shared/bulk.sv.js'
+import * as bulksvjson from '#shared/bulk.svjson.js'
+import * as bulktrunc from '#shared/bulk.trunc.js'
 import * as d3dsv from 'd3-dsv'
 
 export default async function handle_study(req, res) {
